% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{achillesColumns}
\alias{achillesColumns}
\title{Required columns for each of the achilles result tables}
\usage{
achillesColumns(table, version = "5.3", onlyRequired = lifecycle::deprecated())
}
\arguments{
\item{table}{Table for which to see the required columns. One of
"achilles_analysis", "achilles_results", or "achilles_results_dist".}

\item{version}{Version of the OMOP Common Data Model. Supported options are:
"5.3", "5.4", or "5.5".}

\item{onlyRequired}{deprecated.}
}
\value{
Character vector with the column names
}
\description{
Required columns for each of the achilles result tables
}
\examples{
\donttest{
library(omopgenerics)
achillesColumns("achilles_analysis")
achillesColumns("achilles_results")
achillesColumns("achilles_results_dist")
}
}
