% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{achillesTable}
\alias{achillesTable}
\title{Create an achilles table from a cdm_table.}
\usage{
achillesTable(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\value{
An achilles_table object
}
\description{
Create an achilles table from a cdm_table.
}
