% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{addClass}
\alias{addClass}
\title{To add a new class that will go to the beginning.}
\usage{
addClass(x, class)
}
\value{
Same object with the new class added.
}
\description{
To add a new class that will go to the beginning.
}
