% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{argumentsValidation}
\alias{argumentsValidation}
\title{Details on the performed validation per argument.}
\usage{
argumentsValidation(argumentName = NULL, verbose = TRUE)
}
\arguments{
\item{argumentName}{The of the argument to see the details. If NULL all
available arguments are displayed.}

\item{verbose}{Whether to print the details}
}
\value{
A tibble with the details of the validation (invisible).
}
\description{
Details on the performed validation per argument.
}
\examples{
argumentsValidation("cohortId")
argumentsValidation()

}
