% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{assertDoc}
\alias{assertDoc}
\title{Helper for consistent documentation of assertion functions.}
\arguments{
\item{x}{Variable to check.}

\item{length}{Required length. If \code{NULL} length is not checked.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be NULL.}

\item{empty}{Whether it can be empty.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{x}.}

\item{call}{Call argument that will be passed to \code{cli}.}

\item{msg}{Custom error message.}
}
\description{
Helper for consistent documentation of assertion functions.
}
\keyword{internal}
