% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertTable}
\alias{assertTable}
\title{Assert that an object is a table.}
\usage{
assertTable(
  x,
  class = NULL,
  numberColumns = NULL,
  numberRows = NULL,
  columns = character(),
  allowExtraColumns = TRUE,
  null = FALSE,
  empty = TRUE,
  unique = FALSE,
  nm = deparse1(substitute(x), backtick = TRUE),
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{Variable to check.}

\item{class}{A class that the table must have: "tbl", "data.frame", "tbl_sql",
...}

\item{numberColumns}{Number of columns that it has to contain.}

\item{numberRows}{Number of rows that it has to contain.}

\item{columns}{Name of the columns required.}

\item{allowExtraColumns}{Whether extra columns are allowed.}

\item{null}{Whether it can be NULL.}

\item{empty}{Whether it can be empty.}

\item{unique}{Whether it has to contain unique rows.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{x}.}

\item{call}{Call argument that will be passed to \code{cli}.}

\item{msg}{Custom error message.}
}
\description{
Assert that an object is a table.
}
