% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertTrue}
\alias{assertTrue}
\title{Assert that an expression is TRUE.}
\usage{
assertTrue(
  x,
  null = FALSE,
  empty = TRUE,
  nm = deparse1(substitute(x), backtick = TRUE),
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{Expression to check.}

\item{null}{Whether it can be NULL.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{x}.}

\item{call}{Call argument that will be passed to \code{cli}.}

\item{msg}{Custom error message.}
}
\description{
Assert that an expression is TRUE.
}
