% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAttrition.R
\name{attrition.generated_cohort_set}
\alias{attrition.generated_cohort_set}
\title{Get cohort attrition from a generated_cohort_set object.}
\usage{
\method{attrition}{generated_cohort_set}(x)
}
\arguments{
\item{x}{A generated_cohort_set object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a generated_cohort_set object.
}
