% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAttrition.R
\name{attrition}
\alias{attrition}
\title{Get attrition from an object.}
\usage{
attrition(x)
}
\arguments{
\item{x}{x.}
}
\value{
A table with the attrition.
}
\description{
Get attrition from an object.
}
