% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.cdm_cohort}
\alias{bind.cdm_cohort}
\title{Bind multiple cdm_cohort objects.}
\usage{
\method{bind}{cdm_cohort}(...)
}
\arguments{
\item{...}{Multiple cdm_cohort objects.}
}
\value{
New cdm_cohort
}
\description{
Bind multiple cdm_cohort objects.
}
