% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.cdm_reference}
\alias{bind.cdm_reference}
\title{Bind multiple cdm_reference objects.}
\usage{
\method{bind}{cdm_reference}(..., newCdmName = "BINDED CDM")
}
\arguments{
\item{...}{Multiple cdm_reference objects.}
}
\value{
New cdm_reference.
}
\description{
Bind multiple cdm_reference objects.
}
