% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.generated_cohort_set}
\alias{bind.generated_cohort_set}
\title{Bind multiple generated_cohort_set objects.}
\usage{
\method{bind}{generated_cohort_set}(..., cohortName = "bound_cohort")
}
\arguments{
\item{...}{Multiple generated_cohort_set objects.}

\item{cohortName}{Name of the new bound cohort.}
}
\value{
New generated_cohort_set
}
\description{
Bind multiple generated_cohort_set objects.
}
