% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.GeneratedCohortSet}
\alias{bind.GeneratedCohortSet}
\title{Bind multiple GeneratedCohortSet objects.}
\usage{
\method{bind}{GeneratedCohortSet}(...)
}
\arguments{
\item{...}{Multiple generatedCohortSet objects.}
}
\value{
New generatedCohortSet.
}
\description{
Bind multiple GeneratedCohortSet objects.
}
