% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.omop_cohort}
\alias{bind.omop_cohort}
\title{Bind multiple omop_cohort objects.}
\usage{
\method{bind}{omop_cohort}(...)
}
\arguments{
\item{...}{Multiple omop_cohort objects.}
}
\value{
New omop_cohort
}
\description{
Bind multiple omop_cohort objects.
}
