% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bind}
\alias{bind}
\title{Bind two or more objects of the same class.}
\usage{
bind(...)
}
\arguments{
\item{...}{Objects to bind.}
}
\value{
New object.
}
\description{
Bind two or more objects of the same class.
}
