% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{castDoc}
\alias{castDoc}
\title{Helper for consistent documentation of cast arguments.}
\arguments{
\item{cast}{Whether to cast columns to the correct type.}
}
\description{
Helper for consistent documentation of cast arguments.
}
\keyword{internal}
