% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmAssignTableDoc}
\alias{cdmAssignTableDoc}
\title{Helper for consistent documentation of assigning tables to a CDM reference.}
\arguments{
\item{name}{Name where the new table will be assigned.}

\item{value}{Table with the same source as the \code{cdm_reference} object.}
}
\description{
Helper for consistent documentation of assigning tables to a CDM reference.
}
\keyword{internal}
