% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmClasses}
\alias{cdmClasses}
\title{Separate the cdm tables in classes}
\usage{
cdmClasses(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A list of table names, the name of the list indicates the class.
}
\description{
Separate the cdm tables in classes
}
