% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmDoc}
\alias{cdmDoc}
\title{Helper for consistent documentation of \code{cdm}.}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}
}
\description{
Helper for consistent documentation of \code{cdm}.
}
\keyword{internal}
