% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmFromTable}
\alias{cdmFromTable}
\title{Obtain the cdm that a table comes from.}
\usage{
cdmFromTable(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\description{
Obtain the cdm that a table comes from.
}
