% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmIndexDoc}
\alias{cdmIndexDoc}
\title{Helper for consistent documentation of CDM index functions.}
\arguments{
\item{x}{A \verb{<cdm_reference>} or \verb{<cdm_table>} object.}

\item{name}{Name(s) of the CDM table(s).}
}
\description{
Helper for consistent documentation of CDM index functions.
}
\keyword{internal}
