% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmNameDoc}
\alias{cdmNameDoc}
\title{Helper for consistent documentation of CDM names.}
\arguments{
\item{cdmName}{Name of the \verb{<cdm_reference>} object.}
}
\description{
Helper for consistent documentation of CDM names.
}
\keyword{internal}
