% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmOrTableDoc}
\alias{cdmOrTableDoc}
\title{Helper for consistent documentation of cdm reference or table arguments.}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} or \verb{<cdm_table>} object.}
}
\description{
Helper for consistent documentation of cdm reference or table arguments.
}
\keyword{internal}
