% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmReference.tbl}
\alias{cdmReference.tbl}
\title{\code{cdm_reference} objects constructor}
\usage{
\method{cdmReference}{tbl}(cdmTables, cohortTables, cdmName, ...)
}
\arguments{
\item{cdmTables}{List of standard tables in the OMOP Common Data Model.}

\item{cohortTables}{List of tables that contains \code{generated_cohort_set}
objects.}

\item{cdmName}{Name of the cdm.}

\item{...}{For compatibility.}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
