% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{cdmTable}
\alias{cdmTable}
\title{Extract a table from a cdm reference.}
\usage{
cdmTable(table)
}
\arguments{
\item{table}{A table that it is part of a cdm.}
}
\value{
A table from the cdm reference
}
\description{
Extract a table from a cdm reference.
}
