% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmTableDoc}
\alias{cdmTableDoc}
\title{Helper for consistent documentation of \code{cdm_table} objects.}
\arguments{
\item{table}{A \verb{<cdm_table>} object.}
}
\description{
Helper for consistent documentation of \code{cdm_table} objects.
}
\keyword{internal}
