% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodCdmTableFromSource.R
\name{cdmTableFromSource}
\alias{cdmTableFromSource}
\title{This is an internal developer focused function that creates a cdm_table from
a table that shares the source but it is not a cdm_table.
Please use insertTable if you want to insert a table to a cdm_reference
object.}
\usage{
cdmTableFromSource(src, value)
}
\arguments{
\item{src}{A cdm_source object.}

\item{value}{A table that shares source with the cdm_reference object.}
}
\value{
A cdm_table.
}
\description{
This is an internal developer focused function that creates a cdm_table from
a table that shares the source but it is not a cdm_table.
Please use insertTable if you want to insert a table to a cdm_reference
object.
}
