% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMUtilities-package.R
\docType{package}
\name{CDMUtilities-package}
\alias{CDMUtilities}
\alias{CDMUtilities-package}
\title{CDMUtilities: Methods and Utilities for the OMOP Common Data Model}
\description{
This package provides methods and utilities functionalities for the OMOP Common Data Model based packages. This package also mantains the core classes.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
