% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmVersion.cdm_reference}
\alias{cdmVersion.cdm_reference}
\title{Version of a cdm_reference.}
\usage{
\method{cdmVersion}{cdm_reference}(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
Version of the cdm_reference.
}
\description{
Version of a cdm_reference.
}
