% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkElements}
\alias{checkElements}
\title{Check if it contains a list of omop objects.}
\usage{
checkElements(elements)
}
\arguments{
\item{elements}{}
}
\value{
The same object (invisible) or an error if the object does not
fulfill the conditions.
}
\description{
Check if it contains a list of omop objects.
}
\examples{
\donttest{
library(CDMUtilities)

cdm <- mockCdm()

checkElements(list(mockCdm = cdm))
}

}
