% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/checkInput.R
\name{check_input}
\alias{check_input}
\alias{checkInput}
\title{Check the input parameters in OMOP CDM Tools environment}
\usage{
check_input(..., options = list(), call = parent.frame())

checkInput(..., .options = list(), call = parent.frame())
}
\arguments{
\item{...}{Named elements to check. The name will determine the check that is
applied.}

\item{call}{The corresponding function call is retrieved and mentioned in
error messages as the source of the error.}

\item{.options}{Other paramters needed to conduct the checks. It must be a
named list.}
}
\value{
Informative error and warnings messages if the inputs don't pass the
desired checks.
}
\description{
Check the input parameters in OMOP CDM Tools environment
}
\examples{
\donttest{
library(CDMUtilities)
library(dplyr)

# cdm <- mockCdm()
# checkInput(cdm = cdm)
}

}
