% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkListTables}
\alias{checkListTables}
\title{Check if it is a list of tables from the same source}
\usage{
checkListTables(listTables)
}
\arguments{
\item{listTables}{List of tables to check.}
}
\description{
Check if it is a list of tables from the same source
}
