% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkPath}
\alias{checkPath}
\title{Check if it is a valid path}
\usage{
checkPath(path)
}
\arguments{
\item{path}{Path to a folder}
}
\value{
The same object (invisible) or an error if the object does not
fulfill the conditions.
}
\description{
Check if it is a valid path
}
\examples{
\donttest{
checkPath(path = "home/Documents/CDMUtilities")
}

}
