% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkResultsStem}
\alias{checkResultsStem}
\title{Check if it is a valid resultStem}
\usage{
checkResultsStem(resultsStem)
}
\arguments{
\item{resultsStem}{Character string (at least 5 characters) to precede the
exported objects.}
}
\value{
The same object (invisible) or an error if the object does not
fulfill the conditions.
}
\description{
Check if it is a valid resultStem
}
\examples{
\donttest{
checkResultsStem(resultsStem = "results_mock_cdm")
}

}
