% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSnakeCase}
\alias{checkSnakeCase}
\title{Check whether a character vector is snake case.}
\usage{
checkSnakeCase(string, type = "error", call = parent.frame())
}
\arguments{
\item{string}{A character vector to check if it is snake_case.}

\item{type}{Can either be "error" or "warning".}

\item{call}{The corresponding function call is retrieved and mentioned in
error messages as the source of the error.}
}
\value{
Throws an error/warning if the string is not snake_case
}
\description{
Check whether a character vector is snake case.
}
