% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkZip}
\alias{checkZip}
\title{Check if zip should be displayed}
\usage{
checkZip(zip)
}
\arguments{
\item{zip}{Whether a zip file is created with all the generated csv}
}
\value{
The same object (invisible) or an error if the object does not
fulfill the conditions.
}
\description{
Check if zip should be displayed
}
\examples{
\donttest{
checkZip(zip = TRUE)
}

}
