% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cliCallDoc}
\alias{cliCallDoc}
\title{Helper for consistent documentation of call arguments.}
\arguments{
\item{call}{Call argument passed to \code{cli} functions.}
}
\description{
Helper for consistent documentation of call arguments.
}
\keyword{internal}
