% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{codelist}
\alias{codelist}
\title{'codelist' object constructor}
\usage{
codelist(x, name = "codelist")
}
\arguments{
\item{x}{a named list where each element contains either a vector of concept
IDs or a tibble with a column concept_id that contains concept IDs
representing the codelist}

\item{name}{}
}
\value{
A codelist
}
\description{
'codelist' object constructor
}
