% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{codelistDoc}
\alias{codelistDoc}
\title{Helper for consistent documentation of \code{codelist} objects.}
\arguments{
\item{x}{A \verb{<codelist>} object.}
}
\description{
Helper for consistent documentation of \code{codelist} objects.
}
\keyword{internal}
