% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{codelistWithDetailsDoc}
\alias{codelistWithDetailsDoc}
\title{Helper for consistent documentation of \code{codelist_with_details} objects.}
\arguments{
\item{x}{A \verb{<codelist_with_details>} object.}
}
\description{
Helper for consistent documentation of \code{codelist_with_details} objects.
}
\keyword{internal}
