% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{cohortAttrition.generated_cohort_set}
\alias{cohortAttrition.generated_cohort_set}
\title{Get cohort attrition from a generated_cohort_set object.}
\usage{
\method{cohortAttrition}{generated_cohort_set}(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a generated_cohort_set object.
}
