% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{cohortAttrition}
\alias{cohortAttrition}
\title{Get cohort attrition from a omop_cohort object.}
\usage{
cohortAttrition(cohort)
}
\arguments{
\item{cohort}{A omop_cohort object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a omop_cohort object.
}
