% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodCounts.R
\name{cohortCount.generated_cohort_set}
\alias{cohortCount.generated_cohort_set}
\title{Get cohort counts from a generated_cohort_set object.}
\usage{
\method{cohortCount}{generated_cohort_set}(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a generated_cohort_set object.
}
