% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cohortDoc}
\alias{cohortDoc}
\title{Helper for consistent documentation of \code{cohort_table} objects.}
\arguments{
\item{cohort}{A \verb{<cohort_table>} object.}
}
\description{
Helper for consistent documentation of \code{cohort_table} objects.
}
\keyword{internal}
