% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSet.R
\name{cohortSet.generated_cohort_set}
\alias{cohortSet.generated_cohort_set}
\title{Get cohort settings from a generated_cohort_set object.}
\usage{
\method{cohortSet}{generated_cohort_set}(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a generated_cohort_set object.
}
