% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmCohort.R
\name{cohort_set}
\alias{cohort_set}
\alias{cohortSet}
\title{Get cohort settings from a omop_cohort object.}
\usage{
cohort_set(cohort)

cohortSet(cohort)
}
\arguments{
\item{cohort}{A omop_cohort object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a omop_cohort object.
}
