% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortTable.R
\name{cohortTable}
\alias{cohortTable}
\title{\code{cohort_table} objects constructor.}
\usage{
cohortTable(
  table,
  cohortSetRef = attr(table, "cohort_set"),
  cohortAttritionRef = attr(table, "cohort_attrition"),
  overwrite = TRUE,
  .softValidation = FALSE
)
}
\arguments{
\item{table}{cdm_table object with at least: cohort_definition_id,
subject_id, cohort_start_date, cohort_end_date.}

\item{cohortSetRef}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionRef}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{overwrite}{Whether to overwrite a preexisting table with same name.}

\item{.softValidation}{Whether to perform a soft validation of consistency.}
}
\value{
A cohort_table object
}
\description{
\code{cohort_table} objects constructor.
}
