% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cohortValidationChecksDoc}
\alias{cohortValidationChecksDoc}
\title{Helper for consistent documentation of cohort validation checks.}
\arguments{
\item{checkEndAfterStart}{If TRUE a check that all cohort end dates come on
or after cohort start date will be performed.}

\item{checkOverlappingEntries}{If TRUE a check that no individuals have
overlapping cohort entries will be performed.}

\item{checkMissingValues}{If TRUE a check that there are no missing values in
required fields will be performed.}

\item{checkInObservation}{If TRUE a check that cohort entries are within
the individuals observation periods will be performed.}
}
\description{
Helper for consistent documentation of cohort validation checks.
}
\keyword{internal}
