% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{collect.cdm_cohort}
\alias{collect.cdm_cohort}
\title{To collect a \code{cdm_cohort} object.}
\usage{
collect.cdm_cohort(x, ...)
}
\arguments{
\item{x}{\code{cohort_cdm} object.}

\item{...}{Not used (for compatibility).}
}
\description{
To collect a \code{cdm_cohort} object.
}
