% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{collect.generated_cohort_set}
\alias{collect.generated_cohort_set}
\title{To collect a \code{generated_cohort_set} object.}
\usage{
collect.generated_cohort_set(x, ...)
}
\arguments{
\item{x}{\code{generated_cohort_set} object.}

\item{...}{Not used (for compatibility).}
}
\description{
To collect a \code{generated_cohort_set} object.
}
