% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{collect.omop_cohort}
\alias{collect.omop_cohort}
\title{To collect a \code{omop_cohort} object.}
\usage{
\method{collect}{omop_cohort}(x, ...)
}
\arguments{
\item{x}{\code{omop_cohort} object.}

\item{...}{Not used (for compatibility).}
}
\description{
To collect a \code{omop_cohort} object.
}
