% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classComparedResult.R
\name{comparedResult}
\alias{comparedResult}
\title{'compared_results' object constructor}
\usage{
comparedResult(x, name = "compared_result")
}
\arguments{
\item{x}{input must be a tibble}

\item{name}{Name of the compared result object}
}
\value{
A compared_result object
}
\description{
'compared_results' object constructor
}
