% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodComputeCdmTable.R
\name{computeCdmTable}
\alias{computeCdmTable}
\title{Compute a cdm table.}
\usage{
computeCdmTable(x, name)
}
\arguments{
\item{name}{Name of the table to insert.}

\item{cdm}{A cdm reference.}
}
\value{
The table in the cdm reference.
}
\description{
Compute a cdm table.
}
