% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{conceptCdmDoc}
\alias{conceptCdmDoc}
\title{Helper for consistent documentation of \code{cdm} concept validation.}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object. If provided, concept IDs in \code{x} will
be checked against \code{cdm$concept}.}
}
\description{
Helper for consistent documentation of \code{cdm} concept validation.
}
\keyword{internal}
