% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConceptSet.R
\name{conceptSet}
\alias{conceptSet}
\title{'conceptSet' object constructor}
\usage{
conceptSet(x)
}
\arguments{
\item{x}{a named list of tibbles, each of which containing concept set
definitions}
}
\value{
A conceptSet
}
\description{
'conceptSet' object constructor
}
