% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConceptSetExpression.R
\name{conceptSetExpression}
\alias{conceptSetExpression}
\title{'conceptSetExpression' object constructor}
\usage{
conceptSetExpression(x)
}
\arguments{
\item{x}{a named list of tibbles, each of which containing concept set
definitions}
}
\value{
A conceptSetExpression
}
\description{
'conceptSetExpression' object constructor
}
