% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{conceptSetExpressionDoc}
\alias{conceptSetExpressionDoc}
\title{Helper for consistent documentation of \code{concept_set_expression} objects.}
\arguments{
\item{x}{A \verb{<concept_set_expression>} object.}
}
\description{
Helper for consistent documentation of \code{concept_set_expression} objects.
}
\keyword{internal}
