% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{createIndexes}
\alias{createIndexes}
\title{Create the missing indexes}
\usage{
createIndexes(cdm, name = NULL)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{name}{Name(s) of the cdm tables.}
}
\value{
Whether the process was completed successfully.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
