% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnector.R
\name{dbSource}
\alias{dbSource}
\title{Create a source for a OMOP Common Data Model reference in a database.}
\usage{
dbSource(con, writeSchema)
}
\arguments{
\item{con}{Connection to a database.}

\item{writeSchema}{Schema, catalog and prefix of the database that allow us
to write permanent tables.}
}
\description{
Create a source for a OMOP Common Data Model reference in a database.
}
