% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{domainId}
\alias{domainId}
\title{Function to return the domain id column name of a omop table.}
\usage{
domainId(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the domain id column name
}
\description{
Function to return the domain id column name of a omop table.
}
