% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodDropSourceTable.R
\name{dropSourceTable}
\alias{dropSourceTable}
\title{Drop a table from a cdm object.}
\usage{
dropSourceTable(cdm, name)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} or \verb{<cdm_table>} object.}

\item{name}{Name(s) of the table(s) to drop. Tidyselect statements are
supported.}
}
\value{
The table in the cdm reference.
}
\description{
Drop a table from a cdm object.
}
