% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{emptyCodelist}
\alias{emptyCodelist}
\title{Empty \code{codelist} object.}
\usage{
emptyCodelist()
}
\value{
An empty codelist object.
}
\description{
Empty \code{codelist} object.
}
\examples{
emptyCodelist()

}
