% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelistWithDetails.R
\name{emptyCodelistWithDetails}
\alias{emptyCodelistWithDetails}
\title{Empty \code{codelist} object.}
\usage{
emptyCodelistWithDetails()
}
\value{
An empty codelist object.
}
\description{
Empty \code{codelist} object.
}
\examples{
emptyCodelistWithDetails()

}
