% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classComparedResult.R
\name{emptyComparedResult}
\alias{emptyComparedResult}
\title{Empty \code{compared_result} object.}
\usage{
emptyComparedResult()
}
\value{
An empty \code{compared_result} object.
}
\description{
Empty \code{compared_result} object.
}
\examples{
\donttest{
library(omopgenerics)

emptyComparedResult()
}

}
