% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{emptyDoc}
\alias{emptyDoc}
\title{Helper for consistent documentation of empty arguments.}
\arguments{
\item{empty}{Whether it can be empty.}
}
\description{
Helper for consistent documentation of empty arguments.
}
\keyword{internal}
