% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{emptyTableNameDoc}
\alias{emptyTableNameDoc}
\title{Helper for consistent documentation of empty table names.}
\arguments{
\item{name}{Name of the table to create.}
}
\description{
Helper for consistent documentation of empty table names.
}
\keyword{internal}
