% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{endDate}
\alias{endDate}
\title{Function to return the end date column name of a omop table.}
\usage{
endDate(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the end date column name
}
\description{
Function to return the end date column name of a omop table.
}
