% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{estimateTypeChoices}
\alias{estimateTypeChoices}
\title{Choices that can be present in \code{estimate_type} column.}
\usage{
estimateTypeChoices()
}
\value{
A character vector with the options that can be present in
\code{estimate_type} column in the summarised_result objects.
}
\description{
Choices that can be present in \code{estimate_type} column.
}
