% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{existingIndexes}
\alias{existingIndexes}
\title{Existing indexes in a cdm object}
\usage{
existingIndexes(cdm, name)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{name}{Name(s) of the cdm tables.}
}
\value{
A tibble with 3 columns: \code{table_class} class of the table,
\code{table_name} name of the table, and \code{existing_index} index definition.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
