% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{expectedIndexes}
\alias{expectedIndexes}
\title{Expected indexes in a cdm object}
\usage{
expectedIndexes(x, name)
}
\arguments{
\item{x}{A \verb{<cdm_reference>} or \verb{<cdm_table>} object.}

\item{name}{Name(s) of the CDM table(s).}
}
\value{
A tibble with 3 columns: \code{table_class} class of the table,
\code{table_name} name of the table, and \code{expected_index} index definition.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
