% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.cdm_cohort}
\alias{export.cdm_cohort}
\title{Export a cdm_cohort}
\usage{
\method{export}{cdm_cohort}(x)
}
\arguments{
\item{x}{A cdm_cohort object.}
}
\value{
Exported cdm_cohort.
}
\description{
Export a cdm_cohort
}
