% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.cdm_reference}
\alias{export.cdm_reference}
\title{Export a cdm_reference}
\usage{
\method{export}{cdm_reference}(
  x,
  path = here::here(),
  namePrefix = "",
  minCellCount = 5,
  studyId = NULL
)
}
\arguments{
\item{x}{A cdm_reference object.}

\item{path}{Path where to export files.}

\item{namePrefix}{Prefix for the files that are going to be saved.}

\item{minCellCount}{Minimum number of counts to be exported.}

\item{studyId}{Identifier to be added as \code{study_id}.}
}
\value{
Exported cdm_reference.
}
\description{
Export a cdm_reference
}
