% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.generated_cohort_set}
\alias{export.generated_cohort_set}
\title{Export a generated_cohort_set}
\usage{
\method{export}{generated_cohort_set}(x)
}
\arguments{
\item{x}{A generated_cohort_set object.}
}
\value{
Exported generated_cohort_set
}
\description{
Export a generated_cohort_set
}
