% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.omop_cohort}
\alias{export.omop_cohort}
\title{Export a omop_cohort}
\usage{
\method{export}{omop_cohort}(x)
}
\arguments{
\item{x}{A omop_cohort object.}
}
\value{
Exported omop_cohort.
}
\description{
Export a omop_cohort
}
