% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExport.R
\name{export.summarised_result}
\alias{export.summarised_result}
\title{Export a summarised_result object}
\usage{
\method{export}{summarised_result}(x, path, namePrefix = "", minCellCount = 5, resultId = NULL)
}
\arguments{
\item{x}{A summarised_result object.}

\item{path}{Path where to export files.}

\item{namePrefix}{Prefix for the files that are going to be saved.}

\item{minCellCount}{Minimum number of counts to be exported.}

\item{resultId}{Unique identifier for your set of results.}
}
\value{
Exported summarised_result.
}
\description{
Export a summarised_result object
}
