% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCodelistWithDetails.R
\name{exportCodelistWithDetails}
\alias{exportCodelistWithDetails}
\title{Export a codelist with details object.}
\usage{
exportCodelistWithDetails(x, path, type = "json")
}
\arguments{
\item{x}{A \verb{<codelist_with_details>} object.}

\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently 'json' and 'csv' are
supported.}
}
\value{
Files with codelists with details
}
\description{
Export a codelist with details object.
}
