% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{exportCodeSearch}
\alias{exportCodeSearch}
\title{Export a \code{code_search} object into an Excel spreadsheet}
\usage{
exportCodeSearch(codeSearch, file, path = getwd())
}
\arguments{
\item{codeSearch}{A \code{code_search} object.}

\item{file}{Name of the Excel spreadsheet to be created.}

\item{path}{Path where to create the Excel file.}
}
\value{
The \code{code_search} object is exported to a file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
