% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportConceptSetExpression.R
\name{exportConceptSetExpression}
\alias{exportConceptSetExpression}
\title{Export a concept set expression.}
\usage{
exportConceptSetExpression(x, path, type = "json")
}
\arguments{
\item{x}{A \verb{<concept_set_expression>} object.}

\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently 'json' and 'csv' are
supported.}
}
\value{
Files with concept set expressions
}
\description{
Export a concept set expression.
}
