% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{exportFileDoc}
\alias{exportFileDoc}
\title{Helper for consistent documentation of exported files.}
\arguments{
\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently 'json' and 'csv' are
supported.}
}
\description{
Helper for consistent documentation of exported files.
}
\keyword{internal}
