% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportObject.cdm_reference}
\alias{exportObject.cdm_reference}
\title{Export a cdm_reference}
\usage{
\method{exportObject}{cdm_reference}(x)
}
\arguments{
\item{x}{A cdm_reference object.}
}
\value{
Exported cdm_reference.
}
\description{
Export a cdm_reference
}
