% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export list of package results}
\usage{
exportResults(resultList, zipName, outputFolder)
}
\arguments{
\item{resultList}{Named list with results from a darwin package}

\item{zipName}{name to give zip folder}

\item{outputFolder}{directory to save zip folder containing results as a set
of CSV files}
}
\value{
zip folder of results saved in the outputFolder
}
\description{
Export list of package results
}
\examples{
\donttest{
exportResults(
  resultList = list("mtcars" = mtcars),
  zipName = "test",
  outputFolder = tempdir()
)
}
}
