% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldsTable.R
\docType{data}
\name{fieldsTable}
\alias{fieldsTable}
\title{field tables data}
\format{
\subsection{\code{fieldsTable}}{

A data frame with 1000 rows and 5 columns:
\describe{
\item{cdm_table_name}{omop table name}
\item{cdm_field_name}{column name}
\item{is_required}{compulsory column TRUE or FALSE}
\item{cdm_version}{omop cdm version}
\item{type}{omop table type}
...
}
}
}
\source{
\url{https://www.ohdsi.org/web/wiki/doku.php?id=documentation:cdm:standardized_clinical_data_tables}
}
\usage{
fieldsTable
}
\description{
Dataset with all columns and its details
Report ...
}
\keyword{datasets}
