% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterResult}
\alias{filterResult}
\title{Filter a \verb{<summarised_result>} automatically}
\usage{
filterResult(result, ...)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}

\item{...}{Expressions that return a logical value. The columns used in the
expressions can be stored in the settings, strata, group, or additional
columns. \code{filterResult()} looks for columns using this hierarchy: settings,
strata, group, then additional. If multiple expressions are included, they
are combined with the & operator. Only rows for which all conditions evaluate
to TRUE are kept.}
}
\value{
A \verb{<summarised_result>} object with only the rows that satisfy the
specified filters.
}
\description{
Filter a \verb{<summarised_result>} automatically
}
\examples{
library(dplyr)
library(omopgenerics)

x <- tibble(
  "result_id" = 1L,
  "cdm_name" = "eunomia",
  "group_name" = "cohort_name",
  "group_level" = "my_cohort",
  "strata_name" = "sex",
  "strata_level" = "Female",
  "variable_name" = "number subjects",
  "variable_level" = NA_character_,
  "estimate_name" = "count",
  "estimate_type" = "integer",
  "estimate_value" = "100",
  "additional_name" = "overall",
  "additional_level" = "overall"
) |>
  newSummarisedResult(settings = tibble(
    "result_id" = 1L, "analysis" = "overall"
  ))

x |>
  filterResult(cohort_name == "my_cohort", sex == "Female")

}
