% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdm.R
\name{generateClinicalDataTable}
\alias{generateClinicalDataTable}
\title{Function to add a clinical data table in a cdm_reference.}
\usage{
generateClinicalDataTable(
  cdm,
  tableName,
  numberRows,
  seed = NULL,
  cdmVersion = attr(cdm, "cdm_version")
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{tableName}{Name of the table to be created.}

\item{numberRows}{Number of records in the table.}

\item{seed}{Seed for random numbers reproducibility.}

\item{cdmVersion}{Verison of the cdm.}
}
\value{
A cdm_reference object
}
\description{
Function to add a clinical data table in a cdm_reference.
}
