% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdm.R
\name{generateCohortTable}
\alias{generateCohortTable}
\title{Function to add a cohort table in a cdm_reference.}
\usage{
generateCohortTable(
  cdm,
  tableName = "cohort",
  cohortId = 1,
  cohortName = paste0("cohort_", cohortId),
  numberRows = 100,
  seed = NULL,
  cdmVersion = attr(cdm, "cdm_version")
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{tableName}{Name of the cohort to be created.}

\item{cohortId}{Ids of the cohort to generate.}

\item{cohortName}{Names of the cohort to generate.}

\item{numberRows}{Number of records per cohort. It can have a different
number per cohort.}

\item{seed}{Seed for random numbers reproducibility.}

\item{cdmVersion}{Verison of the cdm.}
}
\value{
A cdm_reference object
}
\description{
Function to add a cohort table in a cdm_reference.
}
