% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{generatedCohortSet.tbl}
\alias{generatedCohortSet.tbl}
\title{\code{generated_cohort_set} objects constructor.}
\usage{
\method{generatedCohortSet}{tbl}(
  cohortTable,
  cohortSetTable = NULL,
  cohortAttritionTable = NULL,
  cohortName = "cohort"
)
}
\arguments{
\item{cohortTable}{Table with at least: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{cohortSetTable}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionTable}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{cohortName}{Name of the generated_cohort_set object.}

\item{cohortCountTable}{Table with at least: cohort_definition_id,
number_subjects, number_records.}
}
\value{
A generated_cohort_set object
}
\description{
\code{generated_cohort_set} objects constructor.
}
