% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdm.R
\name{generatePersonObservationPeriod}
\alias{generatePersonObservationPeriod}
\title{Function create the observation period in a cdm reference.}
\usage{
generatePersonObservationPeriod(
  cdm,
  individuals = NULL,
  seed = NULL,
  cdmVersion = attr(cdm, "cdm_version")
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{individuals}{A tibble with the number of individuals, sex, year of
birth, start observation and end observation.}

\item{seed}{Seed for random numbers reproducibility.}

\item{cdmVersion}{Verison of the cdm.}
}
\value{
A cdm_reference object
}
\description{
Function create the observation period in a cdm reference.
}
