% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodGetCdmSource.R
\name{getCdmSource}
\alias{getCdmSource}
\title{Get the cdm source of an object.}
\usage{
getCdmSource(x)
}
\arguments{
\item{x}{Object to get the cdm source from.}
}
\value{
The cdm source of the object.
}
\description{
Get the cdm source of an object.
}
