% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCohortName}
\alias{getCohortName}
\title{Get the cohort name of a certain cohort definition id}
\usage{
getCohortName(cohort, cohortId = NULL)
}
\arguments{
\item{cohort}{A \verb{<cohort_table>} object.}

\item{cohortId}{Cohort definition id of interest. If NULL all cohort ids are
shown.}
}
\value{
Cohort names
}
\description{
Get the cohort name of a certain cohort definition id
}
