% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFieldTables}
\alias{getFieldTables}
\title{Table of cdm field and its format type}
\usage{
getFieldTables()
}
\value{
tibble of cdm field informations
}
\description{
Table of cdm field and its format type
}
