% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{guessCdmVersion}
\alias{guessCdmVersion}
\title{Guess the OMOP CDM version}
\usage{
guessCdmVersion(cdm)
}
\arguments{
\item{cdm}{A cdm reference or a named list of cdm tables.}
}
\value{
A character vector indicating the guessed cdm version, or
\code{NA_character_} if it cannot be guessed.
}
\description{
Guess the OMOP Common Data Model version from the \code{cdm_source} table. If it
cannot be obtained from there, the table and column names in the cdm are
compared with each supported version and the best match is returned. If
there is no unique best match, \code{NA_character_} is returned.
}
\examples{
cdm <- list(
  cdm_source = dplyr::tibble(cdm_version = "v5.4.1")
)

guessCdmVersion(cdm)

}
