% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCodelist.R
\name{importCodelist}
\alias{importCodelist}
\title{Import a codelist.}
\usage{
importCodelist(path, type = "json")
}
\arguments{
\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently 'json' and 'csv' are
supported.}
}
\value{
A codelist
}
\description{
Import a codelist.
}
