% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCodelistWithDetails.R
\name{importCodelistWithDetails}
\alias{importCodelistWithDetails}
\title{Import a codelist with details.}
\usage{
importCodelistWithDetails(path, type = NULL, recursive = FALSE)
}
\arguments{
\item{path}{Path to a file or directory to import.}

\item{type}{Type of files to import. If \code{NULL}, all supported file types are
imported. Currently 'json' and 'csv' are supported.}

\item{recursive}{If TRUE and path is a directory, search for files will
recurse into directories.}
}
\value{
A codelist_with_details object.
}
\description{
Import a codelist with details.
}
