% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{importCodeSearch}
\alias{importCodeSearch}
\title{Import a \code{code_search} object from an Excel spreadsheet}
\usage{
importCodeSearch(path)
}
\arguments{
\item{path}{Path to the Excel spreadsheet.}
}
\value{
A \code{code_search} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
