% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{importFileDoc}
\alias{importFileDoc}
\title{Helper for consistent documentation of imported files.}
\arguments{
\item{path}{Path to a file or directory to import.}

\item{type}{Type of files to import. If \code{NULL}, all supported file types are
imported. Currently 'json' and 'csv' are supported.}

\item{recursive}{If TRUE and path is a directory, search for files will
recurse into directories.}
}
\description{
Helper for consistent documentation of imported files.
}
\keyword{internal}
