% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{inputValidation}
\alias{inputValidation}
\title{Input validation of arguments, arguments will be corrected accordingly.
You can check the possible validated arguments and their validation in
\code{argumentsValidation()}.}
\usage{
inputValidation(..., call = parent.frame(), .options = list())
}
\arguments{
\item{...}{Named arguments to validate.}

\item{call}{Call object passed to cli messages.}

\item{.options}{named list of internal arguments to be passed to the
validation functions}
}
\description{
Input validation of arguments, arguments will be corrected accordingly.
You can check the possible validated arguments and their validation in
\code{argumentsValidation()}.
}
