% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertCdmTable.R
\name{insertCdmTable}
\alias{insertCdmTable}
\title{Insert a table to a cdm object.}
\usage{
insertCdmTable(cdm, name, table)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of the table to insert.}

\item{table}{Table to insert to the cdm.}
}
\value{
The table in the cdm reference.
}
\description{
Insert a table to a cdm object.
}
