% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertFromSource.R
\name{insertFromSource}
\alias{insertFromSource}
\title{Convert a table that is not a cdm_table but have the same original source to
a cdm_table.
This Table is not meant to be used to insert tables in the cdm, please use
insertTable instead.}
\usage{
insertFromSource(cdm, value)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}

\item{value}{A table that shares source with the cdm_reference object.}
}
\value{
A table in the cdm_reference environment
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
