% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertTable.R
\name{insertTable.local_cdm}
\alias{insertTable.local_cdm}
\title{Insert a table to a local cdm object.}
\usage{
\method{insertTable}{local_cdm}(src, name, table)
}
\arguments{
\item{src}{The source of a local cdm.}

\item{name}{Name of the table to insert.}

\item{table}{Table to insert to the cdm.}
}
\value{
The table in the cdm reference.
}
\description{
Insert a table to a local cdm object.
}
