% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/logger.R
\name{link_logger}
\alias{link_logger}
\alias{linkLogger}
\title{Link a logger file to a cdm object}
\usage{
link_logger(cdm, file)

linkLogger(cdm, file)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{file}{A path to an existing file or place to create it.}
}
\value{
The cdm object with the linked logger.
}
\description{
Link a logger file to a cdm object
}
