% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodListSourceTables.R
\name{listSourceTables}
\alias{listSourceTables}
\title{List tables that can be accessed through a cdm object.}
\usage{
listSourceTables(cdm)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}
}
\value{
A character vector with the names of tables.
}
\description{
List tables that can be accessed through a cdm object.
}
