% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodListTables.R
\name{listTables}
\alias{listTables}
\title{List tables that can be accessed though a cdm object.}
\usage{
listTables(cdm)
}
\arguments{
\item{cdm}{A cdm reference or the source of a cdm reference.}
}
\value{
A character vector with the names of tables.
}
\description{
List tables that can be accessed though a cdm object.
}
