% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLocalSource.R
\name{localSource}
\alias{localSource}
\title{Create a source for a local cdm.}
\usage{
localSource(name)
}
\arguments{
\item{name}{Name of the local cdm.}
}
\description{
Create a source for a local cdm.
}
