% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdm.R
\name{mockCdm}
\alias{mockCdm}
\title{It creates a mock cdm_reference.}
\usage{
mockCdm(
  cdmVocabuly,
  cdmName = "MOCK CDM",
  individuals = 10,
  person = NULL,
  observation_period = NULL,
  death = NULL,
  condition_occurrence = NULL,
  drug_exposure = NULL,
  procedure_occurrence = NULL,
  device_exposure = NULL,
  measurement = NULL,
  observation = NULL,
  seed = 1,
  ...
)
}
\arguments{
\item{cdmVocabuly}{A cdm_reference that contains the vocabulary tables.}

\item{cdmName}{Name of the cdm.}

\item{individuals}{Number of individuals in the mock database or table with
the demographics data.}

\item{person}{Person table.}

\item{observation_period}{Observation period table.}

\item{death}{death table.}

\item{condition_occurrence}{Condition occurrence table.}

\item{drug_exposure}{DrugExposure table.}

\item{procedure_occurrence}{Procedure occurrence table.}

\item{device_exposure}{Device exposure table.}

\item{measurement}{Measurement table.}

\item{observation}{Observation table.}

\item{seed}{Random seed.}

\item{...}{Tables to be added to the cdm. If cohorts the attributes will be
updated.}
}
\value{
A cdm reference with the mock tables
}
\description{
It creates a mock cdm_reference.
}
\examples{
\donttest{
library(CDMUtilities)

cdm <- mockCdm()

cdm
}

}
