% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{newAchillesTable}
\alias{newAchillesTable}
\title{Create an achilles table from a cdm_table.}
\usage{
newAchillesTable(table, version = "5.3", cast = FALSE)
}
\arguments{
\item{table}{A cdm_table.}

\item{version}{version of the cdm.}

\item{cast}{Whether to cast columns to the correct type.}
}
\value{
An achilles_table object
}
\description{
Create an achilles table from a cdm_table.
}
