% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmSource.R
\name{newCdmSource}
\alias{newCdmSource}
\title{Create a cdm source object.}
\usage{
newCdmSource(src, sourceType)
}
\arguments{
\item{src}{Source for a cdm object.}

\item{sourceType}{Type of the source object.}
}
\value{
A validated cdm source object.
}
\description{
Create a cdm source object.
}
