% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{newCodelist}
\alias{newCodelist}
\title{'codelist' object constructor}
\usage{
newCodelist(x, cdm = NULL)
}
\arguments{
\item{x}{A named list where each element contains a vector of concept IDs.}

\item{cdm}{A \verb{<cdm_reference>} object. If provided, concept IDs in \code{x} will
be checked against \code{cdm$concept}.}
}
\value{
A codelist object.
}
\description{
'codelist' object constructor
}
