% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelistWithDetails.R
\name{newCodelistWithDetails}
\alias{newCodelistWithDetails}
\title{'codelist' object constructor}
\usage{
newCodelistWithDetails(x, cdm = NULL)
}
\arguments{
\item{x}{A named list where each element contains a tibble with the column
concept_id}

\item{cdm}{A \verb{<cdm_reference>} object. If provided, concept IDs in \code{x} will
be checked against \code{cdm$concept}.}
}
\value{
A codelist object.
}
\description{
'codelist' object constructor
}
