% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{newCodeSearch}
\alias{newCodeSearch}
\title{Create a new \code{code_search} object}
\usage{
newCodeSearch(codes, searchStrategy)
}
\arguments{
\item{codes}{A tibble with concept ids. It must contain "concept_id",
"found_from", "concept_name", "vocabulary_version", "domain_id",
"vocabulary_id", "concept_class_id", "standard_concept", "concept_code",
"valid_start_date", "valid_end_date" and "invalid_reason" as columns.}

\item{searchStrategy}{A tibble with the search strategy used to derive the
codes. It must contain "strategy_id", "strategy_name" and "strategy_value"
as columns.}
}
\value{
A \code{code_search} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
