% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classComparedResult.R
\name{newComparedResult}
\alias{newComparedResult}
\title{'compared_results' object constructor}
\usage{
newComparedResult(x)
}
\arguments{
\item{x}{input must be a tibble}
}
\value{
A compared_result object
}
\description{
'compared_results' object constructor
}
