% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{newGeneratedCohortSet}
\alias{newGeneratedCohortSet}
\title{CdmReference objects constructor.}
\usage{
newGeneratedCohortSet(
  cohortTable,
  cohortSetTable = NULL,
  cohortAttritionTable = NULL,
  cohortCountTable = NULL,
  validate = TRUE
)
}
\arguments{
\item{cohortTable}{Table with at least: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{cohortSetTable}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionTable}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{cohortCountTable}{Table with at least: cohort_definition_id,
number_subjects, number_records}

\item{validate}{Whether to validate the GeneratedCohortSet.}
}
\value{
A GeneratedCohortSet option.
}
\description{
CdmReference objects constructor.
}
