% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmCohort.R
\name{new_omop_cohort}
\alias{new_omop_cohort}
\alias{newOmopCohort}
\title{\code{omop_cohort} objects constructor.}
\usage{
new_omop_cohort(
  cohort_table,
  cohort_set_table = NULL,
  cohort_attrition_table = NULL,
  cohort_count_table = NULL,
  cohort_name = "cohort"
)

newOmopCohort(
  cohortTable,
  cohortSetTable = NULL,
  cohortAttritionTable = NULL,
  cohortCountTable = NULL,
  cohortName = "cohort"
)
}
\arguments{
\item{cohortTable}{Table with at least: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{cohortSetTable}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionTable}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{cohortCountTable}{Table with at least: cohort_definition_id,
number_subjects, number_records.}

\item{cohortName}{Name of the omop_cohort object.}
}
\value{
A omop_cohort object
}
\description{
\code{omop_cohort} objects constructor.
}
