% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classReadOnly.R
\name{newReadOnlySource}
\alias{newReadOnlySource}
\title{A new read-only source for the cdm}
\usage{
newReadOnlySource()
}
\value{
A read-only source in the format of a cdm source.
}
\description{
This source can be used for \code{cdm_reference} objects that should allow reading
existing CDM tables, but should not allow creating, computing, inserting,
reading from, or dropping source tables.
}
\examples{
\donttest{
library(omopgenerics)
newReadOnlySource()
}

}
