% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{obscureGroups}
\alias{obscureGroups}
\title{Obscure group variable}
\usage{
obscureGroups(result, minCellCount, variable, estimateType, group, groupCount)
}
\arguments{
\item{result}{summarised_result object}

\item{minCellCount}{Minimum count of records to report results.}

\item{variable}{Columns to suppress counts.}

\item{estimateType}{Types to suppress counts.}

\item{group}{Columns to group by and suppress.}

\item{groupCount}{Identifier of the group count.}
}
\value{
Table with obscure_group column
}
\description{
Obscure group variable
}
