% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{obscureLinked}
\alias{obscureLinked}
\title{Obscure variables like percentage}
\usage{
obscureLinked(result, linkEstimates, variable)
}
\arguments{
\item{result}{summarised_result object}

\item{linkEstimates}{List of estimates to link.}

\item{variable}{Columns to suppress counts.}
}
\value{
Table with suppressed counts
}
\description{
Obscure variables like percentage
}
