% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{obscureRecords}
\alias{obscureRecords}
\title{Obscure records}
\usage{
obscureRecords(result, minCellCount, variable, estimateType)
}
\arguments{
\item{result}{summarised_result object}

\item{minCellCount}{Minimum count of records to report results.}

\item{variable}{Columns to suppress counts.}

\item{estimateType}{Types to suppress counts.}
}
\value{
Table with suppressed counts
}
\description{
Obscure records
}
