% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{omopCdmVersionDoc}
\alias{omopCdmVersionDoc}
\title{Helper for consistent documentation of OMOP CDM versions.}
\arguments{
\item{version}{Version of the OMOP Common Data Model.}
}
\description{
Helper for consistent documentation of OMOP CDM versions.
}
\keyword{internal}
