% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{omopColumns}
\alias{omopColumns}
\title{Required columns that the standard tables in the OMOP Common Data Model must
have.}
\usage{
omopColumns(
  table,
  field = NULL,
  version = "5.3",
  onlyRequired = lifecycle::deprecated()
)
}
\arguments{
\item{table}{Table to see required columns.}

\item{field}{Name of the specific field.}

\item{version}{Version of the OMOP Common Data Model.}

\item{onlyRequired}{deprecated}
}
\value{
Character vector with the column names
}
\description{
Required columns that the standard tables in the OMOP Common Data Model must
have.
}
\examples{
library(omopgenerics)

omopColumns("person")

}
