% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPGenerics-package.R
\docType{package}
\name{omopgenerics-package}
\alias{omopgenerics}
\alias{omopgenerics-package}
\title{omopgenerics: Methods and Classes for the OMOP Common Data Model}
\description{
Provides definitions of core classes and methods used by analytic pipelines that query the OMOP (Observational Medical Outcomes Partnership) common data model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/omopgenerics/}
  \item Report bugs at \url{https://github.com/darwin-eu/omopgenerics/issues}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
}

Other contributors:
\itemize{
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID}) [contributor]
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID}) [contributor]
  \item Adam Black \email{black@ohdsi.org} (\href{https://orcid.org/0000-0001-5576-8701}{ORCID}) [contributor]
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID}) [contributor]
}

}
\keyword{internal}
