% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{omopTableFields}
\alias{omopTableFields}
\title{Return a table of omop cdm fields informations}
\usage{
omopTableFields(cdmVersion = "5.3")
}
\arguments{
\item{cdmVersion}{cdm version of the omop cdm.}
}
\value{
a tibble contain informations on all the different fields in omop cdm.
}
\description{
Return a table of omop cdm fields informations
}
