% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{omopTables}
\alias{omopTables}
\title{Standard tables that a cdm reference can contain in the OMOP Common Data
Model.}
\usage{
omopTables(version = "5.3")
}
\arguments{
\item{version}{Version of the OMOP Common Data Model. Supported options are:
"5.3", "5.4", or "5.5".}
}
\value{
Standard tables
}
\description{
Standard tables that a cdm reference can contain in the OMOP Common Data
Model.
}
\examples{
library(omopgenerics)

omopTables()

}
