% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPUtilities-package.R
\docType{package}
\name{OMOPUtilities-package}
\alias{OMOPUtilities}
\alias{OMOPUtilities-package}
\title{OMOPUtilities: Methods, Utilities and Classes for the OMOP Common Data Model}
\description{
This package provides methods and utilities functionalities for the OMOP Common Data Model based packages. This package also mantains the core classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu-dev.github.io/OMOPUtilities/}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
