% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{overwriteDoc}
\alias{overwriteDoc}
\title{Helper for consistent documentation of overwrite arguments.}
\arguments{
\item{overwrite}{Whether to overwrite an existing table.}
}
\description{
Helper for consistent documentation of overwrite arguments.
}
\keyword{internal}
