% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodParticipants.R
\name{participants}
\alias{participants}
\title{It returns the participants that contributed to a particular analysis}
\usage{
participants(result, ...)
}
\arguments{
\item{result}{A result object.}

\item{...}{...}
}
\value{
Table with the participants
}
\description{
It returns the participants that contributed to a particular analysis
}
