% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{print.code_search}
\alias{print.code_search}
\title{Print a code search}
\usage{
\method{print}{code_search}(x, ...)
}
\arguments{
\item{x}{A \code{code_search} object.}

\item{...}{For compatibility with the generic.}
}
\value{
Invisibly returns the input.
}
\description{
Print a code search
}
