% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{print.codelist}
\alias{print.codelist}
\title{Print a codelist}
\usage{
\method{print}{codelist}(x, ...)
}
\arguments{
\item{x}{A \verb{<codelist>} object.}

\item{...}{For compatibility; not used.}
}
\value{
Invisibly returns the input
}
\description{
Print a codelist
}
\examples{
codes <- list("disease X" = c(1, 2, 3), "disease Y" = c(4, 5))
codes <- newCodelist(codes)
print(codes)

}
