% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelistWithDetails.R
\name{print.codelist_with_details}
\alias{print.codelist_with_details}
\title{Print a codelist with details}
\usage{
\method{print}{codelist_with_details}(x, ...)
}
\arguments{
\item{x}{A \verb{<codelist_with_details>} object.}

\item{...}{For compatibility; not used.}
}
\value{
Invisibly returns the input
}
\description{
Print a codelist with details
}
\examples{
codes <- list("disease X" = dplyr::tibble(
  concept_id = c(1, 2, 3),
  other = c("a", "b", "c")
))
codes <- newCodelistWithDetails(codes)
print(codes)

}
