% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{randomPrefix}
\alias{randomPrefix}
\title{Create a random prefix for tables, that contains a random sample of lowercase
letters.}
\usage{
randomPrefix(n = 5)
}
\arguments{
\item{n}{Number of characters.}
}
\value{
A random prefix.
}
\description{
Create a random prefix for tables, that contains a random sample of lowercase
letters.
}
