% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodReadTable.R
\name{readTable}
\alias{readTable}
\title{Read a table from a cdm object.}
\usage{
readTable(src, name)
}
\arguments{
\item{src}{A cdm reference or the source of a cdm reference.}

\item{name}{Name of the table to read.}
}
\value{
The table in the cdm reference.
}
\description{
Read a table from a cdm object.
}
