% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{recursiveDoc}
\alias{recursiveDoc}
\title{Helper for consistent documentation of recursive file imports.}
\arguments{
\item{recursive}{If TRUE and path is a directory, search for files will
recurse into directories.}
}
\description{
Helper for consistent documentation of recursive file imports.
}
\keyword{internal}
