% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{requiredAchillesColumns}
\alias{requiredAchillesColumns}
\title{Required columns for achilles result tables}
\usage{
requiredAchillesColumns(table)
}
\arguments{
\item{table}{Either achilles_analysis, achilles_results,
achilles_results_dist}
}
\value{
Names of columns for achilles result tables
}
\description{
Required columns for achilles result tables
}
