% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{requiredOmopCdmColumns}
\alias{requiredOmopCdmColumns}
\title{Required columns that the standard tables in the OMOP Common Data Model must
have.}
\usage{
requiredOmopCdmColumns(table, version = "5.3")
}
\arguments{
\item{table}{Table to see required columns.}

\item{version}{Version of the OMOP Common Data Model.}
}
\value{
Required columns
}
\description{
Required columns that the standard tables in the OMOP Common Data Model must
have.
}
