% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classResolvedConceptSet.R
\name{resolvedConceptSet}
\alias{resolvedConceptSet}
\title{\code{resolved_concept_set} objects constructor.}
\usage{
resolvedConceptSet(cohortTable)
}
\arguments{
\item{codelist}{List with each item containing either a vector of concept IDs
or a tibble with a column called concept_id.}
}
\value{
A resolved_concept_set object
}
\description{
\code{resolved_concept_set} objects constructor.
}
