% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classResultCollection.R
\name{resultCollection}
\alias{resultCollection}
\title{\code{results_collection} objects constructor}
\usage{
resultCollection(...)
}
\arguments{
\item{...}{Tibble as result objects.}
}
\value{
A \code{results_collection} object.
}
\description{
\code{results_collection} objects constructor
}
