% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{resultColumns}
\alias{resultColumns}
\title{Required columns that the result tables must have.}
\usage{
resultColumns(table = "summarised_result", internal = FALSE)
}
\arguments{
\item{table}{Table to see required columns.}

\item{internal}{For compatibility, do not use please.}
}
\value{
Required columns
}
\description{
Required columns that the result tables must have.
}
\examples{
library(omopgenerics)

resultColumns()

}
