% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resultPackageVersion}
\alias{resultPackageVersion}
\title{Check if different package versions are used for a summarised_result object}
\usage{
resultPackageVersion(result)
}
\arguments{
\item{result}{A summarised_result object.}
}
\value{
A summarised_result object.
}
\description{
Check if different package versions are used for a summarised_result object
}
