% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultType.R
\name{resultType}
\alias{resultType}
\title{Get the \code{result_type(s)} defined in a certain package}
\usage{
resultType(pkg = NULL)
}
\arguments{
\item{pkg}{A character vector of the packages of interest}
}
\value{
A named (package) list of the result_type registered in each package.
}
\description{
Get the \code{result_type(s)} defined in a certain package
}
\examples{
library(omopgenerics)

resultType()

}
