% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{searchStrategy}
\alias{searchStrategy}
\title{Get the search strategy used to create a \code{code_search}}
\usage{
searchStrategy(codeSearch)
}
\arguments{
\item{codeSearch}{A \code{code_search} object.}
}
\value{
A search strategy tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
