% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSet.R
\name{set.generated_cohort_set}
\alias{set.generated_cohort_set}
\title{Get cohort settings from a generated_cohort_set object.}
\usage{
\method{set}{generated_cohort_set}(x)
}
\arguments{
\item{x}{A generated_cohort_set object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a generated_cohort_set object.
}
