% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTemporary.R
\name{setTemporary}
\alias{setTemporary}
\title{Set temporary (or premanent) behaviour for compute functions.}
\usage{
setTemporary(intermediateAsTemp = TRUE, cohortAsTemp = FALSE)
}
\arguments{
\item{intermediate}{Whether intermediate tables should be temporary.}

\item{cohort}{Whether cohort tables should be temporary.}
}
\description{
Set temporary (or premanent) behaviour for compute functions.
}
