% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSettings.R
\name{settings.generated_cohort_set}
\alias{settings.generated_cohort_set}
\title{Get cohort settings from a generated_cohort_set object.}
\usage{
\method{settings}{generated_cohort_set}(x)
}
\arguments{
\item{x}{A generated_cohort_set object.}
}
\value{
A table with the details of the cohort settings.
}
\description{
Get cohort settings from a generated_cohort_set object.
}
