% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSettings.R
\name{settings}
\alias{settings}
\title{Get settings from an object.}
\usage{
settings(x)
}
\arguments{
\item{x}{Object}
}
\value{
A table with the settings of the object.
}
\description{
Get settings from an object.
}
