% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sourceConcept}
\alias{sourceConcept}
\title{Function to return the source concept column name of a omop table.}
\usage{
sourceConcept(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the source concept column name
}
\description{
Function to return the source concept column name of a omop table.
}
