% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardConcept}
\alias{standardConcept}
\title{Function to return the standard concept column name of a omop table.}
\usage{
standardConcept(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the standard concept column name
}
\description{
Function to return the standard concept column name of a omop table.
}
