% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{standardOmopCdmTables}
\alias{standardOmopCdmTables}
\title{Standard tables that a cdm reference can contain in the OMOP Common Data
Model.}
\usage{
standardOmopCdmTables(version = "5.3")
}
\arguments{
\item{version}{Version of the OMOP Common Data Model.}
}
\value{
Standard tables
}
\description{
Standard tables that a cdm reference can contain in the OMOP Common Data
Model.
}
