% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{startDate}
\alias{startDate}
\title{Function to return the start date column name of a omop table.}
\usage{
startDate(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the start date column name
}
\description{
Function to return the start date column name of a omop table.
}
