% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{subsetResult}
\alias{subsetResult}
\title{Subset a summaried_result or compared_result object to a certain result_type.}
\usage{
subsetResult(result, resultType)
}
\arguments{
\item{result}{A result object.}

\item{resultType}{A result type identifier.}
}
\value{
A subsetted
}
\description{
Subset a summaried_result or compared_result object to a certain result_type.
}
