% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{summarisedResult}
\alias{summarisedResult}
\title{'summarised_results' object constructor}
\usage{
summarisedResult(x)
}
\arguments{
\item{x}{Table.}
}
\value{
A summarisedResult object
}
\description{
'summarised_results' object constructor
}
