% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{summarisedResultDoc}
\alias{summarisedResultDoc}
\title{Helper for consistent documentation of \code{summarised_result} objects.}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}
}
\description{
Helper for consistent documentation of \code{summarised_result} objects.
}
\keyword{internal}
