% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cdm_source}
\alias{summary.cdm_source}
\title{Summarise a \code{cdm_source} object}
\usage{
\method{summary}{cdm_source}(object, ...)
}
\arguments{
\item{object}{A generated cohort set object.}

\item{...}{For compatibility; not used.}
}
\value{
A list of properties of the \code{cdm_source} object.
}
\description{
Summarise a \code{cdm_source} object
}
\examples{
summary(newLocalSource())

}
