% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary.generated_cohort_set}
\alias{summary.generated_cohort_set}
\title{Summary a generated cohort set}
\usage{
\method{summary}{generated_cohort_set}(object, ...)
}
\arguments{
\item{object}{A generated cohort set object.}

\item{...}{For compatibility (not used).}
}
\description{
Summary a generated cohort set
}
