% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supportedCdmVersions}
\alias{supportedCdmVersions}
\title{Supported OMOP CDM versions}
\format{
A character vector of OMOP CDM version numbers.
}
\usage{
supportedCdmVersions
}
\description{
A character vector containing the OMOP CDM versions supported by
\code{omopgenerics}.
}
\examples{
supportedCdmVersions
}
\keyword{datasets}
