% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSuppress.R
\name{suppress.compared_result}
\alias{suppress.compared_result}
\title{Function to suppress counts in compared_result objects}
\usage{
\method{suppress}{compared_result}(result, minCellCount = 5)
}
\arguments{
\item{result}{compared_result object}

\item{minCellCount}{Minimum count of records to report results.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in compared_result objects
}
