% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{tmpPrefix}
\alias{tmpPrefix}
\title{Create a temporary prefix for tables that contains a unique prefix that
starts with tmp.}
\usage{
tmpPrefix()
}
\value{
A temporary prefix.
}
\description{
Create a temporary prefix for tables that contains a unique prefix that
starts with tmp.
}
\examples{
library(omopgenerics)
tmpPrefix()
}
