% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/utilities.R
\name{to_camel_case}
\alias{to_camel_case}
\alias{toCamelCase}
\title{Convert a character to camelCase}
\usage{
to_camel_case(string)

toCamelCase(string)
}
\arguments{
\item{string}{String to convert to camelCase}
}
\description{
Convert a character to camelCase
}
\examples{
\donttest{
library(OMOPUtilities)
toCamelCase(c("foo bar", "fooBar", "FooBar", "foo_bar", "FOOBAR"))
}

}
