% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toSnakeCase}
\alias{toSnakeCase}
\title{Convert a character vector to snake case}
\usage{
toSnakeCase(x)
}
\arguments{
\item{x}{Character vector to convert}
}
\value{
A snake_case vector
}
\description{
Convert a character vector to snake case
}
\examples{
toSnakeCase("myVariable")

toSnakeCase(c("cohort1", "Cohort22b"))

}
