% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{typeConcept}
\alias{typeConcept}
\title{Function to return the type concept id column name of a omop table.}
\usage{
typeConcept(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the type concept column name
}
\description{
Function to return the type concept id column name of a omop table.
}
