% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{unAggregate}
\alias{unAggregate}
\title{Expand a pair of columns in name-level format to several columns.}
\usage{
unAggregate(x, name = "group_name", level = "group_level", keep = FALSE)
}
\arguments{
\item{x}{Tibble or data.frame.}

\item{name}{Column name of the \code{name} column.}

\item{level}{Column name of the \code{level} column.}

\item{keep}{Whether to keep the original columns.}
}
\value{
A Tibble with the new columns.
}
\description{
Expand a pair of columns in name-level format to several columns.
}
