% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{uniqueId}
\alias{uniqueId}
\title{Get a unique Identifier with a certain number of characters and a prefix.}
\usage{
uniqueId(n = 1, exclude = character(), nChar = 3, prefix = "id_")
}
\arguments{
\item{n}{Number of identifiers.}

\item{exclude}{Columns to exclude.}

\item{nChar}{Number of characters.}

\item{prefix}{A prefix for the identifiers.}
}
\value{
A character vector with n unique identifiers.
}
\description{
Get a unique Identifier with a certain number of characters and a prefix.
}
