% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{uniqueIdentifier}
\alias{uniqueIdentifier}
\title{Function to return the unique identifier id column name of a omop table.}
\usage{
uniqueIdentifier(tableName)
}
\arguments{
\item{tableName}{name of the omop table}
}
\value{
the unique id column name
}
\description{
Function to return the unique identifier id column name of a omop table.
}
