% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{uniqueTableName}
\alias{uniqueTableName}
\title{Create a unique table name}
\usage{
uniqueTableName(prefix = "")
}
\arguments{
\item{prefix}{Prefix for the table names.}
}
\value{
A string that can be used as a dbplyr temp table name
}
\description{
Create a unique table name
}
