% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteGroup.R
\name{uniteGroup}
\alias{uniteGroup}
\title{Unite one or more columns in name-level format.}
\usage{
uniteGroup(x, cols, name = "group_name", level = "group_level", keep = FALSE)
}
\arguments{
\item{x}{Tibble or data.frame.}

\item{cols}{Columns to aggregate.}

\item{name}{Column name of the \code{name} column.}

\item{level}{Column name of the \code{level} column.}

\item{keep}{Whether to keep the original columns.}
}
\value{
A Tibble with the new columns.
}
\description{
Unite one or more columns in name-level format.
}
