% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{unusedDotsDoc}
\alias{unusedDotsDoc}
\title{Helper for consistent documentation of unused dots.}
\arguments{
\item{...}{For compatibility; not used.}
}
\description{
Helper for consistent documentation of unused dots.
}
\keyword{internal}
