% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useAlias.R
\name{useAlias}
\alias{useAlias}
\title{Function to reexport other functions using camelCase or snake_case.}
\usage{
useAlias(
  originalCase = c("camelCase", "snake_case"),
  newCase = c("camelCase", "snake_case")
)
}
\arguments{
\item{originalCase}{Casing of functions to be reexported.}

\item{newCase}{Casing of the new created functions.}
}
\description{
Function to reexport other functions using camelCase or snake_case.
}
