% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateAgeGroupArgument}
\alias{validateAgeGroupArgument}
\title{Validate the ageGroup argument. It must be a list of two integerish numbers
lower age and upper age, both of the must be greater or equal to 0 and lower
age must be lower or equal to the upper age. If not named automatic names
will be given in the output list.}
\usage{
validateAgeGroupArgument(
  ageGroup,
  multipleAgeGroup = TRUE,
  overlap = FALSE,
  null = TRUE,
  empty = TRUE,
  ageGroupName = "age_group",
  nm = deparse1(substitute(ageGroup), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{ageGroup}{age group in a list.}

\item{multipleAgeGroup}{Allow multiple age groups.}

\item{overlap}{allow overlapping ageGroup.}

\item{null}{null age group allowed true or false.}

\item{empty}{Whether it can be empty.}

\item{ageGroupName}{Name of the default age group.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{ageGroup}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
validate ageGroup
}
\description{
Validate the ageGroup argument. It must be a list of two integerish numbers
lower age and upper age, both of the must be greater or equal to 0 and lower
age must be lower or equal to the upper age. If not named automatic names
will be given in the output list.
}
\examples{
validateAgeGroupArgument(list(c(0, 39), c(40, Inf)))

}
