% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{validateCdmCohort.cdm_cohort}
\alias{validateCdmCohort.cdm_cohort}
\title{To assess if an object is a valid \code{cdm_cohort}.}
\usage{
\method{validateCdmCohort}{cdm_cohort}(cohortTable)
}
\arguments{
\item{cohortTable}{Table to validate}
}
\description{
To assess if an object is a valid \code{cdm_cohort}.
}
