% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{validateCdmCohort.omop_cohort}
\alias{validateCdmCohort.omop_cohort}
\title{To assess if an object is a valid \code{omop_cohort}.}
\usage{
\method{validateCdmCohort}{omop_cohort}(cohortTable)
}
\arguments{
\item{cohortTable}{Table to validate}
}
\description{
To assess if an object is a valid \code{omop_cohort}.
}
