% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{validateCdmCohort}
\alias{validateCdmCohort}
\title{To assess if an object is a valid \code{omop_cohort}}
\usage{
validateCdmCohort(cohortTable)
}
\arguments{
\item{cohortTable}{\code{omop_cohort} to validate.}
}
\description{
To assess if an object is a valid \code{omop_cohort}
}
