% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmReference.R
\name{validate_cdm_reference}
\alias{validate_cdm_reference}
\alias{validateCdmReference}
\title{Validate a \code{cdm_reference} object.}
\usage{
validate_cdm_reference(cdm)

validateCdmReference(cdm)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}
}
\description{
Validate a \code{cdm_reference} object.
}
