% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateColumn}
\alias{validateColumn}
\title{Validate whether a variable points to a certain existing column in a table.}
\usage{
validateColumn(
  column,
  x,
  type = c("character", "date", "logical", "numeric", "integer"),
  validation = "error",
  null = FALSE,
  empty = TRUE,
  nm = deparse1(substitute(column), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{column}{Name of a column that you want to check exists in \code{x}
table.}

\item{x}{Table to check if the column exists.}

\item{type}{Type of the column.}

\item{validation}{How to perform validation: "error", "warning".}

\item{null}{Whether \code{NULL} is accepted.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{column}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
the validated name
}
\description{
Validate whether a variable points to a certain existing column in a table.
}
\examples{
x <- dplyr::tibble(a = 1, b = "xxx")

validateColumn("a", x, validation = "warning")
validateColumn("a", x, type = "character", validation = "warning")
validateColumn("a", x, type = "numeric", validation = "warning")
validateColumn("not_existing", x, type = "numeric", validation = "warning")

}
