% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateConceptSetArgument}
\alias{validateConceptSetArgument}
\title{Validate conceptSet argument. It can either be a list, a codelist, a
concept set expression or a codelist with details. The output will always be a
codelist.}
\usage{
validateConceptSetArgument(
  conceptSet,
  cdm = NULL,
  empty = TRUE,
  validation = "error",
  nm = deparse1(substitute(conceptSet), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{conceptSet}{It can be either a named list of concepts or a codelist,
codelist_with_details or concept_set_expression object.}

\item{cdm}{A \verb{<cdm_reference>} object. If provided, concept IDs in the
resulting codelist will be checked against \code{cdm$concept}. It is also needed
if a concept_set_expression with descendants is provided.}

\item{empty}{Whether it can be empty.}

\item{validation}{How to perform validation: "error", "warning".}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{conceptSet}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
A codelist object.
}
\description{
Validate conceptSet argument. It can either be a list, a codelist, a
concept set expression or a codelist with details. The output will always be a
codelist.
}
\examples{
conceptSet <- list(disease_x = c(1L, 2L))
validateConceptSetArgument(conceptSet)

}
