% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{validateGeneratedCohortSet}
\alias{validateGeneratedCohortSet}
\title{To assess if an object is a valid GeneratedCohortSet.}
\usage{
validateGeneratedCohortSet(cohortTable)
}
\arguments{
\item{cohortTable}{Table to validate}
}
\value{
invisible(cohortTable)
}
\description{
To assess if an object is a valid GeneratedCohortSet.
}
\examples{
\donttest{
library(CDMUtilities)

cdm <- mockCdm()

validateGeneratedCohortSet(cdm$cohort1)
}

}
