% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{validateOmopTable}
\alias{validateOmopTable}
\title{Validate an omop_table}
\usage{
validateOmopTable(
  omopTable,
  version = NULL,
  cast = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{omopTable}{An omop_table to check.}

\item{version}{The version of the cdm.}

\item{cast}{Whether to cast columns to the correct type.}

\item{call}{Call argument that will be passed to \code{cli} error message.}
}
\value{
An omop_table object.
}
\description{
Validate an omop_table
}
