% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateResultArguemnt}
\alias{validateResultArguemnt}
\title{validateResultArguemnt}
\usage{
validateResultArguemnt(
  result,
  packageName = NULL,
  validation = "error",
  call = parent.frame()
)
}
\arguments{
\item{result}{summarise result object to validate}

\item{packageName}{name of the package the result from, if NULL ignore.}

\item{validation}{message to return}

\item{call}{}
}
\value{
summarise result object
}
\description{
validateResultArguemnt
}
