% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateSnakeCase}
\alias{validateSnakeCase}
\title{Validate a character vector to ensure it is snake case, it will be converted
to snake case otherwise.}
\usage{
validateSnakeCase(string, type = "message", call = parent.frame())
}
\arguments{
\item{string}{A character vector to validate if it is snake_case.}

\item{type}{Can be either, "message" or "silent".}

\item{call}{The corresponding function call is retrieved and mentioned in
error messages as the source of the error.}
}
\value{
A snake_case character
}
\description{
Validate a character vector to ensure it is snake case, it will be converted
to snake case otherwise.
}
