% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateStrataArgument}
\alias{validateStrataArgument}
\title{To validate a strata list. It makes sure that elements are unique and point
to columns in table.}
\usage{
validateStrataArgument(
  strata,
  table,
  empty = TRUE,
  nm = deparse1(substitute(strata), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{strata}{A list of characters that point to columns in table.}

\item{table}{A table with columns.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{strata}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
The same strata input or an error if the input is incorrect.
}
\description{
To validate a strata list. It makes sure that elements are unique and point
to columns in table.
}
\examples{
strata <- list("age", "sex", c("age", "sex"))
x <- dplyr::tibble(age = 30L, sex = "Female")

validateStrataArgument(strata, x)

}
