% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateSuppressedResult}
\alias{validateSuppressedResult}
\title{Title}
\usage{
validateSuppressedResult(result, minCellCount = 5)
}
\arguments{
\item{result}{The suppressed result to check}

\item{minCellCount}{Minimum count of records used when suppressing}
}
\value{
Warning or message with check result
}
\description{
Title
}
