% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{validationDoc}
\alias{validationDoc}
\title{Helper for consistent documentation of validation mode.}
\arguments{
\item{validation}{How to perform validation: "error", "warning".}
}
\description{
Helper for consistent documentation of validation mode.
}
\keyword{internal}
