% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{writeLog}
\alias{writeLog}
\title{Write message to a log file.}
\usage{
writeLog(message, warn = TRUE)
}
\arguments{
\item{message}{Message to record in the logger.}

\item{warn}{Weather to throw a warning if the logger does not exist.}
}
\value{
Invisible cdm object
}
\description{
Write message to a log file.
}
