% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/logger.R
\name{write_logger}
\alias{write_logger}
\alias{writeLogger}
\title{Write message to a logger file associated to a cdm object.}
\usage{
write_logger(cdm, message, warn = FALSE)

writeLogger(cdm, message, warn = FALSE)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{message}{Message to record in the logger.}

\item{warn}{Weather to throw a warning if the logger does not exist.}
}
\value{
Invisible cdm object
}
\description{
Write message to a logger file associated to a cdm object.
}
