% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{achillesColumns}
\alias{achillesColumns}
\title{Required columns for achilles result tables}
\usage{
achillesColumns(table, version = "5.3")
}
\arguments{
\item{table}{Table to see required columns.}

\item{version}{Version of the OMOP Common Data Model.}
}
\value{
Names of columns for achilles result tables
}
\description{
Required columns for achilles result tables
}
