% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{achillesTables}
\alias{achillesTables}
\title{Names of the tables that contain the results of achilles analyses}
\usage{
achillesTables(version = "5.3")
}
\arguments{
\item{version}{Version of the OMOP Common Data Model.}
}
\value{
Names of the tables that are contain the results from the achilles
analyses
}
\description{
Names of the tables that contain the results of achilles analyses
}
\examples{
\donttest{
library(omopgenerics)
achillesTables()
}
}
