% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{achillesTables}
\alias{achillesTables}
\title{Tables containing the results of achilles analyses}
\usage{
achillesTables(version = "5.3")
}
\arguments{
\item{version}{Version of the OMOP Common Data Model.}
}
\value{
Names of tables returned by achilles analyses
}
\description{
Tables containing the results of achilles analyses
}
