% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{addSettings}
\alias{addSettings}
\title{Add settings columns to a summaries_result object.}
\usage{
addSettings(result, columns = NULL)
}
\arguments{
\item{result}{A summarised_result object.}

\item{columns}{Settings to be added as columns, by default all settings will
be added.}
}
\value{
A summarised_result object with the added setting columns.
}
\description{
Add settings columns to a summaries_result object.
}
