% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/genericChecks.R
\name{assert_character}
\alias{assert_character}
\alias{assertCharacter}
\title{Assert if an object is a character and fulfill certain conditions.}
\usage{
assert_character(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  min_num_character = 0,
  call = parent.frame()
)

assertCharacter(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  minNumCharacter = 0,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli}.}

\item{minNumCharacter}{Minimum number of characters.}
}
\description{
Assert if an object is a character and fulfill certain conditions.
}
