% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertCharacter}
\alias{assertCharacter}
\title{Assert that an object is a character and fulfill certain conditions.}
\usage{
assertCharacter(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  unique = FALSE,
  named = FALSE,
  minNumCharacter = 0,
  call = parent.frame()
)
}
\arguments{
\item{x}{Variable to check.}

\item{length}{Required length.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be NULL.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{minNumCharacter}{Minimum number of characters.}

\item{call}{Call argument that will be passed to \code{cli} error message.}
}
\description{
Assert that an object is a character and fulfill certain conditions.
}
