% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericChecks.R
\name{assertCharacter}
\alias{assertCharacter}
\title{Assert if a character fulfill certain conditions.}
\usage{
assertCharacter(
  x,
  len = NULL,
  missing = FALSE,
  nullOk = FALSE,
  named = FALSE,
  minNumCharacter = 0,
  errorMessage = NULL
)
}
\arguments{
\item{x}{To check.}

\item{len}{Length that has to have.}

\item{missing}{Whether it can contain missing.}

\item{nullOk}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{minNumCharacter}{Minimum number of characters.}

\item{errorMessage}{Error message to display.}
}
\description{
Assert if a character fulfill certain conditions.
}
