% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertChoice}
\alias{assertChoice}
\title{Assert that an object is within a certain oprtions.}
\usage{
assertChoice(
  x,
  choices,
  length = NULL,
  na = FALSE,
  null = FALSE,
  unique = FALSE,
  named = FALSE,
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{Variable to check.}

\item{choices}{Options that x is allowed to be.}

\item{length}{Required length. If \code{NULL} length is not checked.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be NULL.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli} error message.}

\item{msg}{Custom error message.}
}
\description{
Assert that an object is within a certain oprtions.
}
