% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericChecks.R
\name{assertChoice}
\alias{assertChoice}
\title{Assert if an object is a choice and fulfill certain conditions.}
\usage{
assertChoice(
  x,
  choices,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  errorMessage = NULL
)
}
\arguments{
\item{x}{To check.}

\item{choices}{Options that x can be.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{errorMessage}{Error message to display.}
}
\description{
Assert if an object is a choice and fulfill certain conditions.
}
