% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertClass}
\alias{assertClass}
\title{Assert that an object has a certain class.}
\usage{
assertClass(
  x,
  class,
  null = TRUE,
  all = FALSE,
  extra = TRUE,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{class}{Expected class or classes.}

\item{null}{Whether it can be NULL.}

\item{all}{Whether it should have all the classes or only at least one of
them.}

\item{extra}{Whether the object can have extra classes.}

\item{call}{Call argument that will be passed to \code{cli}.}
}
\description{
Assert that an object has a certain class.
}
