% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertClass}
\alias{assertClass}
\title{Assert that an object has a certain class.}
\usage{
assertClass(
  x,
  class,
  length = NULL,
  null = FALSE,
  all = FALSE,
  extra = TRUE,
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{To check.}

\item{class}{Expected class or classes.}

\item{length}{Required length. If \code{NULL} length is not checked.}

\item{null}{Whether it can be NULL.}

\item{all}{Whether it should have all the classes or only at least one of
them.}

\item{extra}{Whether the object can have extra classes.}

\item{call}{Call argument that will be passed to \code{cli}.}

\item{msg}{Custom error message.}
}
\description{
Assert that an object has a certain class.
}
