% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertDate}
\alias{assertDate}
\title{Assert Date}
\usage{
assertDate(x, length, null = FALSE, call = parent.frame(), msg = NULL)
}
\arguments{
\item{x}{Expression to check.}

\item{length}{length of the date vector}

\item{null}{Whether it can be NULL.}

\item{call}{Call argument that will be passed to \code{cli} error message.}

\item{msg}{Custom error message}
}
\value{
x
}
\description{
Assert Date
}
