% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericChecks.R
\name{assertList}
\alias{assertList}
\title{Assert if an object is a list and fulfill certain conditions.}
\usage{
assertList(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  class = NULL,
  errorMessage = NULL
)
}
\arguments{
\item{x}{To check.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{class}{Class that elements must have.}

\item{errorMessage}{Error message to display.}
}
\description{
Assert if an object is a list and fulfill certain conditions.
}
