% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/genericChecks.R
\name{assert_list}
\alias{assert_list}
\alias{assertList}
\title{Assert if an object is a list and fulfill certain conditions.}
\usage{
assert_list(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  class = NULL,
  call = parent.frame()
)

assertList(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  class = NULL,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{class}{Class that elements must have.}

\item{call}{Call argument that will be passed to \code{cli}.}
}
\description{
Assert if an object is a list and fulfill certain conditions.
}
