% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericChecks.R
\name{assertLogical}
\alias{assertLogical}
\title{Assert if an object is a logical and fulfill certain conditions.}
\usage{
assertLogical(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli}.}
}
\description{
Assert if an object is a logical and fulfill certain conditions.
}
