% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertLogical}
\alias{assertLogical}
\title{Assert that an object is a logical.}
\usage{
assertLogical(
  x,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{Variable to check.}

\item{length}{Required length.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be NULL.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli} error message.}

\item{msg}{Custom error message.}
}
\description{
Assert that an object is a logical.
}
