% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertNumeric}
\alias{assertNumeric}
\title{Assert that an object is a numeric.}
\usage{
assertNumeric(
  x,
  integerish = FALSE,
  min = -Inf,
  max = Inf,
  length = NULL,
  na = FALSE,
  null = FALSE,
  unique = FALSE,
  named = FALSE,
  call = parent.frame(),
  msg = NULL
)
}
\arguments{
\item{x}{Variable to check.}

\item{integerish}{Whether it has to be an integer}

\item{min}{Minimum value that the object can be.}

\item{max}{Maximum value that the object can be.}

\item{length}{Required length. If \code{NULL} length is not checked.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be NULL.}

\item{unique}{Whether it has to contain unique elements.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli} error message.}

\item{msg}{Custom error message.}
}
\description{
Assert that an object is a numeric.
}
