% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericChecks.R
\name{assertNumeric}
\alias{assertNumeric}
\title{Assert if an object is a numeric and fulfill certain conditions.}
\usage{
assertNumeric(
  x,
  integerish = FALSE,
  min = -Inf,
  max = Inf,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  errorMessage = NULL
)
}
\arguments{
\item{x}{To check.}

\item{integerish}{Whether elements must be integerish.}

\item{min}{Lower bound.}

\item{max}{Upper bound.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{errorMessage}{Error message to display.}
}
\description{
Assert if an object is a numeric and fulfill certain conditions.
}
