% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/genericChecks.R
\name{assert_numeric}
\alias{assert_numeric}
\alias{assertNumeric}
\title{Assert if an object is a numeric and fulfill certain conditions.}
\usage{
assert_numeric(
  x,
  integerish = FALSE,
  min = -Inf,
  max = Inf,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  call = parent.frame()
)

assertNumeric(
  x,
  integerish = FALSE,
  min = -Inf,
  max = Inf,
  length = NULL,
  na = FALSE,
  null = FALSE,
  named = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{x}{To check.}

\item{integerish}{Whether elements must be integerish.}

\item{min}{Lower bound.}

\item{max}{Upper bound.}

\item{length}{Length that has to have.}

\item{na}{Whether it can contain NA values.}

\item{null}{Whether it can be null.}

\item{named}{Whether it has to be named.}

\item{call}{Call argument that will be passed to \code{cli}.}
}
\description{
Assert if an object is a numeric and fulfill certain conditions.
}
