% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertTable}
\alias{assertTable}
\title{Assert that an object is a table.}
\usage{
assertTable(
  x,
  numberColumns = NULL,
  numberRows = NULL,
  columns = NULL,
  allowExtraColumns = TRUE,
  null = FALSE,
  distinct = FALSE,
  requireLocal = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{x}{Variable to check.}

\item{numberColumns}{Number of columns that it has to contain.}

\item{numberRows}{Number of rows that it has to contain.}

\item{columns}{Name of the columns required.}

\item{allowExtraColumns}{Whether extra columns are allowed.}

\item{null}{Whether it can be NULL.}

\item{distinct}{Whether it has to contain distinct rows.}

\item{requireLocal}{Whether the table has to be a data.frame.}

\item{call}{Call argument that will be passed to \code{cli} error message.}
}
\description{
Assert that an object is a table.
}
