% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAttrition.R
\name{attrition.cohort_table}
\alias{attrition.cohort_table}
\title{Get cohort attrition from a cohort_table object.}
\usage{
\method{attrition}{cohort_table}(x)
}
\arguments{
\item{x}{A cohort_table object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a cohort_table object.
}
