% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAttrition.R
\name{attrition.cohort_table}
\alias{attrition.cohort_table}
\title{Get cohort attrition from a cohort_table object.}
\usage{
\method{attrition}{cohort_table}(x)
}
\arguments{
\item{x}{A cohort_table}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a cohort_table object.
}
\examples{
 \donttest{
library(omopgenerics)
library(dplyr, warn.conflicts = FALSE)

person <- tibble(
  person_id = 1L, gender_concept_id = 0L, year_of_birth = 1990L,
  race_concept_id = 0L, ethnicity_concept_id = 0,
  month_of_birth = NA_integer_, day_of_birth = NA_integer_,
  birth_datetime = as.Date(NA_character_), location_id = 0L,
  provider_id = 0L, care_site_id = 0L, person_source_value = NA_character_,
  gender_source_value = NA_character_,
  gender_source_concept_id = NA_integer_, race_source_value = NA_character_,
  race_source_concept_id = NA_integer_,
  ethnicity_source_value = NA_character_,
  ethnicity_source_concept_id = NA_integer_
)
observation_period <- tibble(
  observation_period_id = 1L, person_id = 1L,
  observation_period_start_date = as.Date("2000-01-01"),
  observation_period_end_date = as.Date("2023-12-31"),
  period_type_concept_id = 0L
)
cohort <- tibble(
  cohort_definition_id = c(1L, 1L, 1L, 2L),
  subject_id = 1L,
  cohort_start_date = as.Date(c("2020-01-01", "2021-01-01", "2022-01-01", "2022-01-01")),
  cohort_end_date = as.Date(c("2020-01-01", "2021-01-01", "2022-01-01", "2022-01-01")),
)
cdm <- cdmFromTables(
  tables = list("person" = person, "observation_period" = observation_period),
  cdmName = "my_example_cdm",
  cohortTables = list("cohort1" = cohort)
)

attrition(cdm$cohort1)
}
}
