% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodBind.R
\name{bind.cohort_table}
\alias{bind.cohort_table}
\title{Bind cohort_table}
\usage{
\method{bind}{cohort_table}(..., name)
}
\arguments{
\item{...}{Generated cohort set objects to bind. At least two must be
provided.}

\item{name}{Name of the new generated cohort set.}
}
\value{
The cdm object with the new generated cohort set.
}
\description{
Bind cohort_table
}
