% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodBind.R
\name{bind.cohort_table}
\alias{bind.cohort_table}
\title{Bind two or more cohort tables}
\usage{
\method{bind}{cohort_table}(..., name)
}
\arguments{
\item{...}{Generated cohort set objects to bind. At least two must be
provided.}

\item{name}{Name of the new generated cohort set.}
}
\value{
The cdm object with a new generated cohort set containing all
of the cohorts passed.
}
\description{
Bind two or more cohort tables
}
\examples{
library(omopgenerics)
library(dplyr, warn.conflicts = FALSE)

cohort1 <- tibble(
  cohort_definition_id = 1L,
  subject_id = 1:3L,
  cohort_start_date = as.Date("2010-01-01"),
  cohort_end_date = as.Date("2010-01-05")
)
cohort2 <- tibble(
  cohort_definition_id = c(2L, 2L, 3L, 3L, 3L),
  subject_id = c(1L, 2L, 3L, 1L, 2L),
  cohort_start_date = as.Date("2010-01-01"),
  cohort_end_date = as.Date("2010-01-05")
)
cdm <- cdmFromTables(
  tables = list(
    "person" = tibble(
      person_id = c(1L, 2L, 3L), gender_concept_id = 0L, year_of_birth = 1990L,
      race_concept_id = 0L, ethnicity_concept_id = 0,
      month_of_birth = NA_integer_, day_of_birth = NA_integer_,
      birth_datetime = as.Date(NA_character_), location_id = 0L,
      provider_id = 0L, care_site_id = 0L, person_source_value = NA_character_,
      gender_source_value = NA_character_,
      gender_source_concept_id = NA_integer_, race_source_value = NA_character_,
      race_source_concept_id = NA_integer_,
      ethnicity_source_value = NA_character_,
      ethnicity_source_concept_id = NA_integer_
    ),
    "observation_period" = tibble(
      observation_period_id = 1:3L, person_id = 1:3L,
      observation_period_start_date = as.Date("2000-01-01"),
      observation_period_end_date = as.Date("2023-12-31"),
      period_type_concept_id = 0L
    )
  ),
  cdmName = "mock",
  cohortTables = list("cohort1" = cohort1, "cohort2" = cohort2)
)

cdm <- bind(cdm$cohort1, cdm$cohort2, name = "cohort3")
settings(cdm$cohort3)
cdm$cohort3

}
