% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodBind.R
\name{bind.generated_cohort_set}
\alias{bind.generated_cohort_set}
\title{Bind generated_cohort_set}
\usage{
\method{bind}{generated_cohort_set}(..., name)
}
\arguments{
\item{...}{Generated cohort set objects to bind. At least two must be
provided.}

\item{name}{Name of the new generated cohort set.}
}
\value{
The cdm object with the new generated cohort set.
}
\description{
Bind generated_cohort_set
}
