% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodBind.R
\name{bind.generated_cohort_set}
\alias{bind.generated_cohort_set}
\title{Bind generated_cohort_set}
\usage{
\method{bind}{generated_cohort_set}(..., name = NULL)
}
\arguments{
\item{...}{Generated cohort set objects to bind. At least two must be
provided.}

\item{name}{Name of the new generated cohort set. If not provided The name of
the first cohort is used instead.}
}
\value{
New generated cohort set
}
\description{
Bind generated_cohort_set
}
