% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodBind.R
\name{bind.summarised_result}
\alias{bind.summarised_result}
\title{Bind two or summarised_result objects}
\usage{
\method{bind}{summarised_result}(...)
}
\arguments{
\item{...}{summarised_result objects}
}
\value{
A summarised_result object the merged objects.
}
\description{
Bind two or summarised_result objects
}
\examples{
library(omopgenerics)
library(dplyr, warn.conflicts = FALSE)

cdm <- cdmFromTables(
  tables = list(
    "person" = tibble(
      person_id = c(1L, 2L, 3L), gender_concept_id = 0L, year_of_birth = 1990L,
      race_concept_id = 0L, ethnicity_concept_id = 0,
      month_of_birth = NA_integer_, day_of_birth = NA_integer_,
      birth_datetime = as.Date(NA_character_), location_id = 0L,
      provider_id = 0L, care_site_id = 0L, person_source_value = NA_character_,
      gender_source_value = NA_character_,
      gender_source_concept_id = NA_integer_, race_source_value = NA_character_,
      race_source_concept_id = NA_integer_,
      ethnicity_source_value = NA_character_,
      ethnicity_source_concept_id = NA_integer_
    ),
    "observation_period" = tibble(
      observation_period_id = 1:3L, person_id = 1:3L,
      observation_period_start_date = as.Date("2000-01-01"),
      observation_period_end_date = as.Date("2023-12-31"),
      period_type_concept_id = 0L
    )
  ),
  cdmName = "mock",
  cohortTables = list("cohort1" = tibble(
    cohort_definition_id = 1L,
    subject_id = 1:3L,
    cohort_start_date = as.Date("2010-01-01"),
    cohort_end_date = as.Date("2010-01-05")
  ))
)

result1 <- summary(cdm)
result2 <- summary(cdm$cohort1)

mergedResult <- bind(result1, result2)
mergedResult

}
