% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromLocal.R
\name{cdmFromTables}
\alias{cdmFromTables}
\title{\code{cdm_reference} objects constructor}
\usage{
cdmFromTables(tables, cdmName, cohortTables = list())
}
\arguments{
\item{cdmName}{Name of the cdm object.}

\item{cohortTables}{List of tables that contains \code{generated_cohort_set}
objects.}

\item{cdmTables}{List of standard tables in the OMOP Common Data Model.}

\item{achillesTables}{List of tables that contain achilles.}

\item{otherTables}{List of tables that contain unespecified tables in the cdm.}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
