% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromTables.R
\name{cdmFromTables}
\alias{cdmFromTables}
\title{\code{cdm_reference} objects constructor}
\usage{
cdmFromTables(tables, cdmName, cohortTables = list())
}
\arguments{
\item{tables}{List of tables to be part of the cdm object.}

\item{cdmName}{Name of the cdm object.}

\item{cohortTables}{List of tables that contains cohort, cohort_set and
cohort_attrition can be provided as attributes.}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
