% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromTables.R
\name{cdmFromTables}
\alias{cdmFromTables}
\title{Create a cdm object from local tables}
\usage{
cdmFromTables(tables, cdmName, cohortTables = list(), cdmVersion = NULL)
}
\arguments{
\item{tables}{List of tables to be part of the cdm object.}

\item{cdmName}{Name of the cdm object.}

\item{cohortTables}{List of tables that contains cohort, cohort_set and
cohort_attrition can be provided as attributes.}

\item{cdmVersion}{Version of the cdm_reference}
}
\value{
A \code{cdm_reference} object.
}
\description{
Create a cdm object from local tables
}
