% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmName}
\alias{cdmName}
\title{Name of a cdm_reference.}
\usage{
cdmName(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
Name of the cdm_reference.
}
\description{
Name of a cdm_reference.
}
