% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmReference.R
\name{cdm_name}
\alias{cdm_name}
\alias{cdmName}
\title{Name of a cdm_reference.}
\usage{
cdm_name(cdm)

cdmName(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
Name of the cdm_reference.
}
\description{
Name of a cdm_reference.
}
