% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmReference}
\alias{cdmReference}
\title{Obtain the cdm_reference that a table comes from.}
\usage{
cdmReference(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\description{
Obtain the cdm_reference that a table comes from.
}
