% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmReference}
\alias{cdmReference}
\title{\code{cdm_reference} objects constructor}
\usage{
cdmReference(cdmTables)
}
\arguments{
\item{cdmTables}{List of standard tables in the OMOP Common Data Model.}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
