% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmReference}
\alias{cdmReference}
\title{\code{cdm_reference} objects constructor}
\usage{
cdmReference(cdmTables, cdmName, cdmVersion)
}
\arguments{
\item{cdmTables}{List of tables that contains a reference to an OMOP Common
Data Model.}

\item{cdmName}{Name of the cdm.}

\item{cdmVersion}{Version of the cdm ("5.3" or "5.4").}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
