% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{cdmReference}
\alias{cdmReference}
\title{Get the \code{cdm_reference} of a \code{cdm_table}.}
\usage{
cdmReference(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\value{
A cdm_reference.
}
\description{
Get the \code{cdm_reference} of a \code{cdm_table}.
}
