% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmSource}
\alias{cdmSource}
\title{Get the source of a cdm_reference.}
\usage{
cdmSource(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A cdm_source object.
}
\description{
Get the source of a cdm_reference.
}
