% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmSource.R
\name{cdmSource}
\alias{cdmSource}
\title{Create a cdm source object.}
\usage{
cdmSource(src, sourceType)
}
\arguments{
\item{src}{Source to a cdm object.}

\item{sourceType}{Type of the source object.}
}
\value{
A validated cdm source object.
}
\description{
Create a cdm source object.
}
