% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmSourceType}
\alias{cdmSourceType}
\title{Get the source type of a cdm_reference object.}
\usage{
cdmSourceType(cdm)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}
}
\value{
A character vector with the type of source of the cdm_reference
object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\donttest{
library(omopgenerics)
library(dplyr, warn.conflicts = FALSE)

cdm <- cdmFromTables(
  tables = list(
    "person" = tibble(
      person_id = c(1, 2, 3), gender_concept_id = 0, year_of_birth = 1990,
      race_concept_id = 0, ethnicity_concept_id = 0
    ),
    "observation_period" = tibble(
      observation_period_id = 1:3, person_id = 1:3,
      observation_period_start_date = as.Date("2000-01-01"),
      observation_period_end_date = as.Date("2023-12-31"),
      period_type_concept_id = 0
    )
  ),
  cdmName = "mock"
)

cdmSourceType(cdm)
}
}
