% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmSourceType}
\alias{cdmSourceType}
\title{Get the source type of a cdm_reference object.}
\usage{
cdmSourceType(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A character vector with the type of source of the cdm_reference
object.
}
\description{
Get the source type of a cdm_reference object.
}
\examples{
\donttest{
library(omopgenerics)
cdmTables <- list(
  "person" = dplyr::tibble(
    person_id = 1, gender_concept_id = 0, year_of_birth = 1990,
    race_concept_id = 0, ethnicity_concept_id = 0
  ) |>
    newCdmTable(newLocalSource(), "person"),
  "observation_period" = dplyr::tibble(
    observation_period_id = 1, person_id = 1,
    observation_period_start_date = as.Date("2000-01-01"),
    observation_period_end_date = as.Date("2025-12-31"),
    period_type_concept_id = 0
  ) |>
    newCdmTable(newLocalSource(), "observation_period")
)
cdm <- newCdmReference(tables = cdmTables, cdmName = "mock")

cdmSourceType(cdm = cdm)
}
}
