% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmSourceType}
\alias{cdmSourceType}
\title{Get the source type of a cdm_reference object.}
\usage{
cdmSourceType(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A character vector with the type of source of the cdm_reference
object.
}
\description{
Get the source type of a cdm_reference object.
}
