% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{cdmTable}
\alias{cdmTable}
\title{Extract a table from a cdm reference.}
\usage{
cdmTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm reference object.}

\item{name}{The name of the element that you want to extract.}
}
\value{
A table from the cdm reference
}
\description{
Extract a table from a cdm reference.
}
