% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{cdmTable}
\alias{cdmTable}
\title{Create an cdm table.}
\usage{
cdmTable(table, src, name)
}
\arguments{
\item{table}{A table that is part of a cdm.}

\item{src}{The source of the table.}

\item{name}{The name of the table.}
}
\value{
A cdm_table object
}
\description{
Create an cdm table.
}
