% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cdmVersion}
\alias{cdmVersion}
\title{Get the version of a cdm_reference.}
\usage{
cdmVersion(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
Version of the cdm_reference.
}
\description{
Get the version of a cdm_reference.
}
