% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmVersionArgumentDoc}
\alias{cdmVersionArgumentDoc}
\title{Helper for consistent documentation of CDM version arguments.}
\arguments{
\item{cdmVersion}{Version of the OMOP Common Data Model. Supported versions
are "5.3" and "5.4".}
}
\description{
Helper for consistent documentation of CDM version arguments.
}
\keyword{internal}
