% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmCohort.R
\name{cohort_attrition}
\alias{cohort_attrition}
\alias{cohortAttrition}
\title{Get cohort attrition from a omop_cohort object.}
\usage{
cohort_attrition(cohort)

cohortAttrition(cohort)
}
\arguments{
\item{cohort}{A omop_cohort object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a omop_cohort object.
}
