% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAttributes.R
\name{cohortAttrition}
\alias{cohortAttrition}
\title{Get cohort attrition from a GeneratedCohortSet object.}
\usage{
cohortAttrition(cohort)
}
\arguments{
\item{cohort}{A GeneratedCohortSet object.}
}
\value{
A table with the attrition.
}
\description{
Get cohort attrition from a GeneratedCohortSet object.
}
