% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{cohortColumns}
\alias{cohortColumns}
\title{Required columns for a generated cohort set.}
\usage{
cohortColumns(table, version = "5.3")
}
\arguments{
\item{table}{Either \code{cohort}, \code{cohort_set} or \code{cohort_attrition}}

\item{version}{Version of the OMOP Common Data Model. Supported options are:
"5.3", "5.4", or "5.5".}
}
\value{
Character vector with the column names

Required columns
}
\description{
Required columns for a generated cohort set.
}
\examples{
\donttest{
library(omopgenerics)
cohortColumns("cohort")
}
}
