% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmCohort.R
\name{cohort_count}
\alias{cohort_count}
\alias{cohortCount}
\title{Get cohort counts from a omop_cohort object.}
\usage{
cohort_count(cohort)

cohortCount(cohort)
}
\arguments{
\item{cohort}{A omop_cohort object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a omop_cohort object.
}
