% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAttributes.R
\name{cohortCount}
\alias{cohortCount}
\title{Get cohort counts from a GeneratedCohortSet object.}
\usage{
cohortCount(cohort)
}
\arguments{
\item{cohort}{A GeneratedCohortSet object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a GeneratedCohortSet object.
}
