% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{cohortCount}
\alias{cohortCount}
\title{Get cohort counts from a generated_cohort_set object.}
\usage{
cohortCount(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a generated_cohort_set object.
}
