% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{cohortCount}
\alias{cohortCount}
\title{Get cohort counts from a omop_cohort object.}
\usage{
cohortCount(cohort)
}
\arguments{
\item{cohort}{A omop_cohort object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a omop_cohort object.
}
