% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortCount.R
\name{cohortCount}
\alias{cohortCount}
\title{Get cohort counts from a cohort_table object.}
\usage{
cohortCount(cohort)
}
\arguments{
\item{cohort}{A cohort_table object.}
}
\value{
A table with the counts.
}
\description{
Get cohort counts from a cohort_table object.
}
