% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmCohort.R
\name{cohortSet}
\alias{cohortSet}
\title{Get cohort settings from a cdm_cohort object.}
\usage{
cohortSet(cohort)
}
\arguments{
\item{cohort}{A cdm_cohort object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a cdm_cohort object.
}
