% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSet.R
\name{cohortSet}
\alias{cohortSet}
\title{Get cohort settings from a generated_cohort_set object.}
\usage{
cohortSet(cohort)
}
\arguments{
\item{cohort}{A generated_cohort_set object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a generated_cohort_set object.
}
