% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAttributes.R
\name{cohortSet}
\alias{cohortSet}
\title{Get cohort settings from a GeneratedCohortSet object.}
\usage{
cohortSet(cohort)
}
\arguments{
\item{cohort}{A GeneratedCohortSet object.}
}
\value{
A table with the details of the cohort set.
}
\description{
Get cohort settings from a GeneratedCohortSet object.
}
