% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineStrata.R
\name{combineStrata}
\alias{combineStrata}
\title{Provide all combinations of strata levels.}
\usage{
combineStrata(levels)
}
\arguments{
\item{levels}{Vector of all strata levels to combine.}
}
\value{
A vector of all combinations of strata.
}
\description{
Provide all combinations of strata levels.
}
