% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compareOmopTableFields}
\alias{compareOmopTableFields}
\title{Compare the fields of two different OMOP CDM versions}
\usage{
compareOmopTableFields(
  cdmVersionReference = "5.3",
  cdmVersionComparator = "5.4"
)
}
\arguments{
\item{cdmVersionReference}{An OMOP CDM version, it can be
"5.3", "5.4", or "5.5".}

\item{cdmVersionComparator}{An OMOP CDM version, it can be
"5.3", "5.4", or "5.5".}
}
\value{
A tibble with two columns: \code{field} and \code{change}.
}
\description{
Compare the fields of two different OMOP CDM versions
}
\examples{
library(omopgenerics)

compareOmopTableFields("5.3", "5.4")

}
