% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute.cdm_table}
\alias{compute.cdm_table}
\title{Store results in a table.}
\usage{
\method{compute}{cdm_table}(x, name = NULL, temporary = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{x}{Table in the cdm.}

\item{name}{Name to store the table with.}

\item{temporary}{Whether to store table temporarily (TRUE) or permanently
(FALSE).}

\item{overwrite}{Whether to overwrite previously existing table with name
same.}

\item{...}{For compatibility (not used).}
}
\value{
Reference to a table in the cdm
}
\description{
Store results in a table.
}
