% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodComputeTable.R
\name{computeTable}
\alias{computeTable}
\title{Compute a cdm table.}
\usage{
computeTable(x, name = NA_character_)
}
\arguments{
\item{name}{Name of the table to insert.}

\item{cdm}{A cdm reference.}
}
\value{
The table in the cdm reference.
}
\description{
Compute a cdm table.
}
