% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{createIndexes}
\alias{createIndexes}
\title{Create the missing indexes}
\usage{
createIndexes(x, name = NULL)
}
\arguments{
\item{x}{A \verb{<cdm_reference>} or \verb{<cdm_table>} object.}

\item{name}{Name(s) of the CDM table(s).}
}
\value{
Whether the process was completed successfully.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
