% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{createTableIndex}
\alias{createTableIndex}
\title{Create a table index}
\usage{
createTableIndex(table, index)
}
\arguments{
\item{table}{A cdm_table object.}

\item{index}{Index to be created.}
}
\value{
Whether the index could be created
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
