% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodDropTable.R
\name{dropTable}
\alias{dropTable}
\title{Drop a table from a cdm object.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
dropTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name(s) of the table(s) to drop Tidyselect statements are
supported.}
}
\value{
The cdm reference.
}
\description{
Drop a table from a cdm object.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
