% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodDropTable.R
\name{dropTable}
\alias{dropTable}
\title{Drop a table from a cdm object.}
\usage{
dropTable(src, name)
}
\arguments{
\item{src}{A cdm reference or the source of a cdm reference.}

\item{name}{Name(s) of the table(s) to insert. Tidyselect statments are
supported.}
}
\value{
The table in the cdm reference.
}
\description{
Drop a table from a cdm object.
}
