% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{emptyAchillesTable}
\alias{emptyAchillesTable}
\title{Create an empty achilles table}
\usage{
emptyAchillesTable(name, cdm)
}
\arguments{
\item{name}{Name of the table to create.}

\item{cdm}{A cdm_reference to create the table.}
}
\value{
The cdm_reference with an achilles empty table
}
\description{
Create an empty achilles table
}
