% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{emptyAchillesTable}
\alias{emptyAchillesTable}
\title{Create an empty achilles table}
\usage{
emptyAchillesTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm_reference to create the table.}

\item{name}{Name of the table to create.}
}
\value{
The cdm_reference with an achilles empty table
}
\description{
Create an empty achilles table
}
\examples{
\donttest{
library(omopgenerics)
cdm <- emptyCdmReference("my_example_cdm")
emptyAchillesTable(cdm = cdm, name = "achilles_results" )
}
}
