% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{emptyCdmReference}
\alias{emptyCdmReference}
\title{Create an empty cdm_reference}
\usage{
emptyCdmReference(cdmName, cdmVersion = NULL)
}
\arguments{
\item{cdmName}{Name of the \verb{<cdm_reference>} object.}

\item{cdmVersion}{Version of the OMOP Common Data Model. Supported options
are: "5.3", "5.4", or "5.5".}
}
\value{
An empty cdm_reference
}
\description{
Create an empty cdm_reference
}
\examples{
\donttest{
library(omopgenerics)
emptyCdmReference(cdmName = "my_example_cdm")
}
}
