% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{emptyCdmReference}
\alias{emptyCdmReference}
\title{Create an empty cdm_reference}
\usage{
emptyCdmReference(cdmName, cdmVersion = NULL)
}
\arguments{
\item{cdmName}{Name of the cdm_reference}

\item{cdmVersion}{Version of the cdm_reference}
}
\value{
An empty cdm_reference
}
\description{
Create an empty cdm_reference
}
