% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortTable.R
\name{emptyCohortTable}
\alias{emptyCohortTable}
\title{Create an empty cohort_table object}
\usage{
emptyCohortTable(name, cdm)
}
\arguments{
\item{name}{Name of the table to create.}

\item{cdm}{A cdm_reference to create the table.}
}
\value{
The cdm_reference with an empty cohort table
}
\description{
Create an empty cohort_table object
}
