% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConceptSetExpression.R
\name{emptyConceptSetExpression}
\alias{emptyConceptSetExpression}
\title{Empty \code{concept_set_expression} object.}
\usage{
emptyConceptSetExpression()
}
\value{
An empty concept_set_expression object.
}
\description{
Empty \code{concept_set_expression} object.
}
\examples{
emptyConceptSetExpression()

}
