% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{emptyOmopTable}
\alias{emptyOmopTable}
\title{Create an empty omop table}
\usage{
emptyOmopTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm_reference to create the table.}

\item{name}{Name of the table to create.}
}
\value{
The cdm_reference with an empty cohort table
}
\description{
Create an empty omop table
}
