% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{emptySummarisedResult}
\alias{emptySummarisedResult}
\title{Empty \code{summarised_result} object.}
\usage{
emptySummarisedResult()
}
\value{
An empty \code{summarised_result} object.
}
\description{
Empty \code{summarised_result} object.
}
\examples{
library(omopgenerics)

emptySummarisedResult()

}
