% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.cdm_reference}
\alias{export.cdm_reference}
\title{Export a cdm_reference}
\usage{
\method{export}{cdm_reference}(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
Exported cdm_reference.
}
\description{
Export a cdm_reference
}
