% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.GeneratedCohortSet}
\alias{export.GeneratedCohortSet}
\title{Export a GeneratedCohortSet.}
\usage{
\method{export}{GeneratedCohortSet}(x)
}
\arguments{
\item{x}{A generatedCohortSet object.}
}
\value{
Exported generatedCohortSet.
}
\description{
Export a GeneratedCohortSet.
}
