% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExport.R
\name{export}
\alias{export}
\title{Export an object.}
\usage{
export(x, path, namePrefix = "", minCellCount = 5, resultId = NULL)
}
\arguments{
\item{x}{Object to export.}

\item{path}{Path where to export files.}

\item{namePrefix}{Prefix for the files that are going to be saved.}

\item{minCellCount}{Minimum number of counts to be exported.}

\item{resultId}{Unique identifier for your set of results.}
}
\value{
Exported object in tbl format.
}
\description{
Export an object.
}
