% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCodelist.R
\name{exportCodelist}
\alias{exportCodelist}
\title{Export a codelist object.}
\usage{
exportCodelist(x, path, type = "json")
}
\arguments{
\item{x}{A codelist}

\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently only "json" is supported.}
}
\value{
Files with codelists
}
\description{
Export a codelist object.
}
