% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export list of results}
\usage{
exportResults(
  ...,
  path = here::here(),
  resultsStem = "results",
  studyId = NULL,
  zip = TRUE
)
}
\arguments{
\item{...}{Named list of results.}

\item{path}{Path to save the results.}

\item{resultsStem}{Stem to save the results csv files. Name of the zip file
if exported.}

\item{zip}{Whether or not to create a zip file with all the results.}
}
\description{
Export list of results
}
