% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummarisedResult.R
\name{exportSummarisedResult}
\alias{exportSummarisedResult}
\title{Export a summarised_result object to a CSV file.}
\usage{
exportSummarisedResult(
  ...,
  minCellCount = 5,
  fileName = "results_{cdm_name}_{date}.csv",
  path = getwd(),
  logFile = getOption("omopgenerics.logFile"),
  logSqlPath = getOption("omopgenerics.log_sql_path"),
  logExplainPath = lifecycle::deprecated()
)
}
\arguments{
\item{...}{A set of summarised_result objects.}

\item{minCellCount}{Minimum count for suppression purposes.}

\item{fileName}{Name of the file that will be created. Use \{cdm_name\} to
refer to the cdmName of the objects and \{date\} to add the export date.}

\item{path}{Path where the CSV file should be created. It is ignored if
fileName is a full file name with a path included.}

\item{logFile}{Path to the log file to export.}

\item{logSqlPath}{Path to the folder that contains the sql logs to export.}

\item{logExplainPath}{deprecated.}
}
\description{
Export a summarised_result object to a CSV file.
}
