% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummarisedResult.R
\name{exportSummarisedResult}
\alias{exportSummarisedResult}
\title{Export a summarised_result object to a csv file.}
\usage{
exportSummarisedResult(
  ...,
  minCellCount = 5,
  fileName = "results_{cdm_name}_{date}.csv",
  path = getwd(),
  temp = FALSE
)
}
\arguments{
\item{...}{A set of summarised_result objects.}

\item{minCellCount}{Minimum count for suppression purposes.}

\item{fileName}{Name of the file that will be created. Use \{cdm_name\} to
refer to the cdmName of the objects and \{date\} to add the export date.}

\item{path}{Path where to create the csv file.}

\item{temp}{whether you are saving a temp file, default as FALSE}
}
\description{
Export a summarised_result object to a csv file.
}
