% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterSettings}
\alias{filterSettings}
\title{Filter a \verb{<summarised_result>} using the settings}
\usage{
filterSettings(result, ...)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}

\item{...}{Expressions that return a logical value (columns in settings are
used to evaluate the expression), and are defined in terms of the variables
in .data. If multiple expressions are included, they are combined with the &
operator. Only rows for which all conditions evaluate to TRUE are kept.}
}
\value{
A \verb{<summarised_result>} object with only the result_id rows that
satisfy the specified settings.
}
\description{
Filter a \verb{<summarised_result>} using the settings
}
\examples{
library(dplyr)
library(omopgenerics)

x <- tibble(
  "result_id" = as.integer(c(1, 2)),
  "cdm_name" = c("cprd", "eunomia"),
  "group_name" = "cohort_name",
  "group_level" = "my_cohort",
  "strata_name" = "sex",
  "strata_level" = "male",
  "variable_name" = "Age group",
  "variable_level" = "10 to 50",
  "estimate_name" = "count",
  "estimate_type" = "numeric",
  "estimate_value" = "5",
  "additional_name" = "overall",
  "additional_level" = "overall"
) |>
  newSummarisedResult(settings = tibble(
    "result_id" = c(1, 2), "custom" = c("A", "B")
  ))

x

x |> filterSettings(custom == "A")

}
