% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCohortId}
\alias{getCohortId}
\title{Get the cohort definition id of a certain name}
\usage{
getCohortId(cohort, cohortName = NULL)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortName}{Names of the cohort of interest. If NULL all cohort names
are shown.}
}
\value{
Cohort definition ids
}
\description{
Get the cohort definition id of a certain name
}
