% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCohortName}
\alias{getCohortName}
\title{Get the cohort name of a certain cohort definition id}
\usage{
getCohortName(cohort, cohortId)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortId}{Cohort deifnition id of interest.}
}
\value{
Cohort names
}
\description{
Get the cohort name of a certain cohort definition id
}
