% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPersonIdentifier}
\alias{getPersonIdentifier}
\title{Get the person identifier of a column (whether subject_id or person_id), it
will throw an error if it contains both.}
\usage{
getPersonIdentifier(x)
}
\arguments{
\item{x}{A table.}
}
\value{
Person identifier column.
}
\description{
Get the person identifier of a column (whether subject_id or person_id), it
will throw an error if it contains both.
}
