% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importConceptSetExpression.R
\name{importConceptSetExpression}
\alias{importConceptSetExpression}
\title{Import a concept set expression.}
\usage{
importConceptSetExpression(path, type = "json")
}
\arguments{
\item{path}{Path to where files will be created.}

\item{type}{Type of files to export. Currently 'json' and 'csv' are
supported.}
}
\value{
A concept set expression
}
\description{
Import a concept set expression.
}
