% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSummarisedResult.R
\name{importSummarisedResult}
\alias{importSummarisedResult}
\title{Import a set of summarised results.}
\usage{
importSummarisedResult(path, recursive = FALSE, ...)
}
\arguments{
\item{path}{Path to directory with CSV files containing summarised results or
to a specific CSV file with a summarised result.}

\item{recursive}{If TRUE and path is a directory, search for files will
recurse into directories.}

\item{...}{Passed to \code{readr::read_csv}.}
}
\value{
A summarised result
}
\description{
Import a set of summarised results.
}
