% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSummarisedResult.R
\name{importSummarisedResult}
\alias{importSummarisedResult}
\title{Import a set of summarised results.}
\usage{
importSummarisedResult(path)
}
\arguments{
\item{path}{Path to directory with CSV files containing summarised results or
to a specific CSV file with a summarised result.}
}
\value{
A summarised result
}
\description{
Import a set of summarised results.
}
