% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertFromSource.R
\name{insertFromSource}
\alias{insertFromSource}
\title{Convert a table that is not a cdm_table but have the same original source to
a cdm_table,}
\usage{
insertFromSource(cdm, value)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{value}{A table that shares source with the cdm_reference object.}
}
\description{
Convert a table that is not a cdm_table but have the same original source to
a cdm_table,
}
