% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertTable.R
\name{insertTable}
\alias{insertTable}
\title{Insert a table to a cdm object.}
\usage{
insertTable(src, name, table)
}
\arguments{
\item{src}{A cdm reference or the source of a cdm reference.}

\item{name}{Name of the table to insert.}

\item{table}{Table to insert to the cdm.}
}
\value{
The table in the cdm reference.
}
\description{
Insert a table to a cdm object.
}
