% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInsertTable.R
\name{insertTable}
\alias{insertTable}
\title{Insert a table to a cdm object.}
\usage{
insertTable(cdm, name, table, overwrite = TRUE)
}
\arguments{
\item{cdm}{A cdm reference or the source of a cdm reference.}

\item{name}{Name of the table to insert.}

\item{table}{Table to insert to the cdm.}

\item{overwrite}{Whether to overwrite an existent table.}
}
\value{
The cdm reference.
}
\description{
Insert a table to a cdm object.
}
