% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isTableEmpty}
\alias{isTableEmpty}
\title{Check if a table is empty or not}
\usage{
isTableEmpty(table)
}
\arguments{
\item{table}{a table}
}
\value{
Boolean to indicate if a cdm_table is empty (TRUE or FALSE).
}
\description{
Check if a table is empty or not
}
