% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVocabulary.R
\name{mockVocabulary}
\alias{mockVocabulary}
\title{It creates a mock database with the vocabulary.}
\usage{
mockVocabulary(
  concept = NULL,
  vocabulary = NULL,
  domain = NULL,
  concept_class = NULL,
  concept_relationship = NULL,
  concept_synonym = NULL,
  concept_ancestor = NULL,
  source_to_concept_map = NULL,
  drug_strength = NULL,
  cdm_version = "5.3"
)
}
\arguments{
\item{concept}{Concept table.}

\item{vocabulary}{Vocabulary table}

\item{domain}{Doamin table.}

\item{concept_class}{Concept_class table.}

\item{concept_relationship}{Concept_relationship table.}

\item{concept_synonym}{Concept_synonym table.}

\item{concept_ancestor}{Concept_ancestor table.}

\item{drug_strength}{Drug_strength table.}

\item{soource_to_concept_map}{Source_to_concept_map table.}
}
\value{
A cdm reference with the vocabulary mock tables
}
\description{
It creates a mock database with the vocabulary.
}
\examples{
\donttest{
library(CDMUtilities)

cdm <- mockVocabulary()

cdm
}

}
