% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{newCdmReference}
\alias{newCdmReference}
\title{CdmReference objects constructor}
\usage{
newCdmReference(cdmTables, cdmName, cdmVersion, validate = TRUE)
}
\arguments{
\item{cdmTables}{List of tables (tbl_sql, tbl or data.frame) that contains a
reference to an OMOP Common data model.}

\item{cdmName}{Name of the cdm.}

\item{cdmVersion}{Version of the cdm ("5.3" or "5.4").}

\item{validate}{Whether to validate the cdm object.}
}
\value{
A CdmReference object.
}
\description{
CdmReference objects constructor
}
