% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmReference.R
\name{new_cdm_reference}
\alias{new_cdm_reference}
\alias{newCdmReference}
\title{\code{cdm_reference} objects constructor}
\usage{
new_cdm_reference(cdm_tables, cdm_name, cdm_version)

newCdmReference(cdmTables, cdmName, cdmVersion)
}
\arguments{
\item{cdmTables}{List of tables (tbl_sql, tbl or data.frame) that contains a
reference to an OMOP Common data model.}

\item{cdmName}{Name of the cdm.}

\item{cdmVersion}{Version of the cdm ("5.3" or "5.4").}
}
\value{
A \code{cdm_reference} object.
}
\description{
\code{cdm_reference} objects constructor
}
