% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmReference.R
\name{newCdmReference}
\alias{newCdmReference}
\title{CdmReference objects constructor}
\usage{
newCdmReference(cdmTables, cdmName)
}
\arguments{
\item{cdmTables}{List of tables (tbl_sql, tbl or data.frame) that contains a
reference to an OMOP Common data model.}

\item{cdmName}{Name of the cdm.}
}
\value{
A cdm reference
}
\description{
CdmReference objects constructor
}
