% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{newCodelist}
\alias{newCodelist}
\title{'codelist' object constructor}
\usage{
newCodelist(x)
}
\arguments{
\item{x}{a named list where each element contains a vector of concept
IDs}
}
\value{
A codelist
}
\description{
'codelist' object constructor
}
