% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortTable.R
\name{newCohortTable}
\alias{newCohortTable}
\title{\code{cohort_table} objects constructor.}
\usage{
newCohortTable(
  table,
  cohortSetRef = attr(table, "cohort_set"),
  cohortAttritionRef = attr(table, "cohort_attrition"),
  .softValidation = FALSE
)
}
\arguments{
\item{table}{cdm_table object with at least: cohort_definition_id,
subject_id, cohort_start_date, cohort_end_date.}

\item{cohortSetRef}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionRef}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{.softValidation}{Whether to perform a soft validation of consistency.
If set to FALSE four additional checks will be performed: 1) a check that
cohort end date is not before cohort start date,  2) a check that there
are no missing values in required columns, 3) a check that cohort duration is
all within observation period, and 4) that there are no overlapping
cohort entries}
}
\value{
A cohort_table object
}
\description{
\code{cohort_table} objects constructor.
}
