% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConceptSetExpression.R
\name{newConceptSetExpression}
\alias{newConceptSetExpression}
\title{'concept_set_expression' object constructor}
\usage{
newConceptSetExpression(x, cdm = NULL)
}
\arguments{
\item{x}{a named list of tibbles, each of which containing concept set
definitions}

\item{cdm}{A \verb{<cdm_reference>} object. If provided, concept IDs in \code{x} will
be checked against \code{cdm$concept}.}
}
\value{
A concept_set_expression
}
\description{
'concept_set_expression' object constructor
}
