% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConceptSetExpression.R
\name{newConceptSetExpression}
\alias{newConceptSetExpression}
\title{'concept_set_expression' object constructor}
\usage{
newConceptSetExpression(x)
}
\arguments{
\item{x}{a named list of tibbles, each of which containing concept set
definitions}
}
\value{
A concept_set_expression
}
\description{
'concept_set_expression' object constructor
}
