% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{newGeneratedCohortSet}
\alias{newGeneratedCohortSet}
\title{\code{generated_cohort_set} objects constructor helper.}
\usage{
newGeneratedCohortSet(
  cohortRef,
  cohortSetRef = NULL,
  cohortAttritionRef = NULL,
  cohortName = "cohort",
  ...
)
}
\arguments{
\item{cohortRef}{Table with at least: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{cohortSetRef}{Table with at least: cohort_definition_id, cohort_name.
If NULL, it will be automatically populated.}

\item{cohortAttritionRef}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records. If NULL, it will be automatically populated.}

\item{cohortName}{Name of the generated_cohort_set object.}

\item{...}{Other arguments for the method.}
}
\value{
A generated_cohort_set object
}
\description{
\code{generated_cohort_set} objects constructor helper.
}
