% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratedCohortSet.R
\name{newGeneratedCohortSet}
\alias{newGeneratedCohortSet}
\title{CdmReference objects constructor.}
\usage{
newGeneratedCohortSet(
  cohortRef,
  cohortSetRef,
  cohortAttritionRef,
  cohortCountRef,
  cdm
)
}
\arguments{
\item{cohortRef}{Table with at least: cohort_definition_id, subject_id,
cohort_start_date, cohort_end_date.}

\item{cohortSetRef}{Table with at least: cohort_definition_id, cohort_name}

\item{cohortAttritionRef}{Table with at least: cohort_definition_id,
number_subjects, number_records, reason_id, reason, excluded_subjects,
excluded_records.}

\item{cohortCountRef}{Table with at least: cohort_definition_id,
number_subjects, number_records}

\item{cdm}{A cdm reference.}
}
\value{
A cdm reference
}
\description{
CdmReference objects constructor.
}
