% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromTables.R
\name{newLocalSource}
\alias{newLocalSource}
\title{A new local source for the cdm}
\usage{
newLocalSource()
}
\value{
A list in the format of a cdm source
}
\description{
A new local source for the cdm
}
\examples{
\donttest{
library(omopgenerics)
newLocalSource()
}

}
