% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{newOmopTable}
\alias{newOmopTable}
\title{Create an omop table from a cdm table.}
\usage{
newOmopTable(table, version = "5.3", cast = FALSE)
}
\arguments{
\item{table}{A \verb{<cdm_table>} object.}

\item{version}{Version of the OMOP Common Data Model.}

\item{cast}{Whether to cast columns to the correct type.}
}
\value{
An omop_table object
}
\description{
Create an omop table from a cdm table.
}
