% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{newOmopTable}
\alias{newOmopTable}
\title{Create an omop table from a cdm table.}
\usage{
newOmopTable(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\value{
An omop_table object
}
\description{
Create an omop table from a cdm table.
}
