% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{newSummarisedResult}
\alias{newSummarisedResult}
\title{'summarised_results' object constructor}
\usage{
newSummarisedResult(x, settings = attr(x, "settings"))
}
\arguments{
\item{x}{Table.}

\item{settings}{Settings for the summarised_result object.}
}
\value{
A \code{summarised_result} object
}
\description{
'summarised_results' object constructor
}
