% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopDataFolder.R
\name{omopDataFolder}
\alias{omopDataFolder}
\title{Check or set the OMOP_DATA_FOLDER where the OMOP related data is stored.}
\usage{
omopDataFolder(path = NULL)
}
\arguments{
\item{path}{Path to a folder to store the OMOP related data. If NULL the
current \code{OMOP_DATA_FOLDER} is returned.}
}
\value{
The OMOP data folder.
}
\description{
Check or set the OMOP_DATA_FOLDER where the OMOP related data is stored.
}
\examples{
\donttest{
omopDataFolder()
omopDataFolder(file.path(tempdir(), "OMOP_DATA"))
omopDataFolder()
}

}
