% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPGenerics-package.R
\docType{package}
\name{OMOPGenerics-package}
\alias{OMOPGenerics}
\alias{OMOPGenerics-package}
\title{OMOPGenerics: Methods and Classes for the OMOP Common Data Model}
\description{
This package define classes and provide methods for the OMOP Common Data Model based packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu-dev.github.io/OMOPGenerics/}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
