% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPGenerics-package.R
\docType{package}
\name{omopgenerics-package}
\alias{omopgenerics}
\alias{omopgenerics-package}
\title{omopgenerics: Methods and Classes for the OMOP Common Data Model}
\description{
The omopgenerics package provides definitions of core classes and methods used by analytic pipelines that query the OMOP common data model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu-dev.github.io/omopgenerics/}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
