% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{participants}
\alias{participants}
\title{It returns the participants that contributed to a particular analysis}
\usage{
participants(...)
}
\arguments{
\item{...}{summarised_result object}
}
\value{
Table with the participants
}
\description{
It returns the participants that contributed to a particular analysis
}
