% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodReadTable.R
\name{readTable}
\alias{readTable}
\title{Read a table from the cdm_source and add it to to the cdm.}
\usage{
readTable(cdm, name)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name of a table to read in the cdm_source space.}
}
\value{
A cdm_reference with new table.
}
\description{
Read a table from the cdm_source and add it to to the cdm.
}
