% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordCohortAttrition.R
\name{recordCohortAttrition}
\alias{recordCohortAttrition}
\title{Update cohort attrition.}
\usage{
recordCohortAttrition(cohort, reason, cohortId = NULL)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{reason}{A character string.}

\item{cohortId}{Cohort definition id of the cohort to update attrition. If
NULL all cohort_definition_id are updated.}
}
\value{
cohort_table with updated attrition.
}
\description{
Update cohort attrition.
}
