% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodTidy.R, R/overwriteClasses.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{generics}{\code{\link[generics]{tidy}}}
}}

