% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resultPackageVersion}
\alias{resultPackageVersion}
\title{Check if different packages version are used for summarise_results object}
\usage{
resultPackageVersion(result)
}
\arguments{
\item{result}{a summarised results object}
}
\value{
a summarised results object
}
\description{
Check if different packages version are used for summarise_results object
}
