% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTemporary.R
\name{setTemporary}
\alias{setTemporary}
\title{Set temporary (or permanent) behavior for compute functions.}
\usage{
setTemporary(intermediateAsTemp = TRUE, cohortAsTemp = FALSE)
}
\arguments{
\item{intermediateAsTemp}{Whether intermediate tables should be temporary.}

\item{cohortAsTemp}{Whether cohort tables should be temporary.}
}
\description{
Set temporary (or permanent) behavior for compute functions.
}
\examples{
\donttest{
library(CDMUtilities)

setTemporary(intermediateAsTemp = TRUE, cohortAsTemp = FALSE)
}

}
