% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSettings.R
\name{settings.cohort_table}
\alias{settings.cohort_table}
\title{Get cohort settings from a cohort_table object.}
\usage{
\method{settings}{cohort_table}(x)
}
\arguments{
\item{x}{A cohort_table object.}
}
\value{
A table with the details of the cohort settings.
}
\description{
Get cohort settings from a cohort_table object.
}
