% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSettings.R
\name{settings.summarised_result}
\alias{settings.summarised_result}
\title{Get settings from a summarised_result object.}
\usage{
\method{settings}{summarised_result}(x)
}
\arguments{
\item{x}{A summarised_result object.}
}
\value{
A table with the settings.
}
\description{
Get settings from a summarised_result object.
}
