% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSettings.R
\name{settings.summarised_result}
\alias{settings.summarised_result}
\title{Get settings from a summarised_result object.}
\usage{
\method{settings}{summarised_result}(x)
}
\arguments{
\item{x}{A summarised_result object.}
}
\value{
A table with the settings.
}
\description{
Get settings from a summarised_result object.
}
\examples{
library(omopgenerics)
library(dplyr, warn.conflicts = FALSE)

person <- tibble(
  person_id = 1L, gender_concept_id = 0L, year_of_birth = 1990L,
  race_concept_id = 0L, ethnicity_concept_id = 0,
  month_of_birth = NA_integer_, day_of_birth = NA_integer_,
  birth_datetime = as.Date(NA_character_), location_id = 0L,
  provider_id = 0L, care_site_id = 0L, person_source_value = NA_character_,
  gender_source_value = NA_character_,
  gender_source_concept_id = NA_integer_, race_source_value = NA_character_,
  race_source_concept_id = NA_integer_,
  ethnicity_source_value = NA_character_,
  ethnicity_source_concept_id = NA_integer_
)
observation_period <- tibble(
  observation_period_id = 1L, person_id = 1L,
  observation_period_start_date = as.Date("2000-01-01"),
  observation_period_end_date = as.Date("2023-12-31"),
  period_type_concept_id = 0L
)
cohort <- tibble(
  cohort_definition_id = 1L,
  subject_id = 1L,
  cohort_start_date = as.Date("2010-01-01"),
  cohort_end_date = as.Date("2012-01-01")
)
cdm <- cdmFromTables(
  tables = list("person" = person, "observation_period" = observation_period),
  cdmName = "test",
  cohortTables = list("my_cohort" = cohort)
)

result <- summary(cdm$my_cohort)

settings(result)

}
