% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmSource.R
\name{sourceType}
\alias{sourceType}
\title{Get the source type of a cdm_source.}
\usage{
sourceType(x)
}
\arguments{
\item{cdmSource}{A cdm_source object.}
}
\value{
A character vector that defines the type of cdm_source.
}
\description{
Get the source type of a cdm_source.
}
