% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmSource.R
\name{sourceType}
\alias{sourceType}
\title{Get the source type of an object.}
\usage{
sourceType(x)
}
\arguments{
\item{x}{Object to know the source type.}
}
\value{
A character vector that defines the type of cdm_source.
}
\description{
Get the source type of an object.
}
