% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{statusIndexes}
\alias{statusIndexes}
\title{Status of the indexes}
\usage{
statusIndexes(cdm, name = NULL)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{name}{Name(s) of the cdm tables.}
}
\value{
A tibble with 3 columns: \code{table_class} class of the table,
\code{table_name} name of the table, \code{index} index definition, and \code{index_status}
status of the index, either: 'missing', 'extra', 'present'.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
