% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{[[<-.cdm_reference}
\alias{[[<-.cdm_reference}
\title{Assign an table to a cdm reference.}
\usage{
\method{[[}{cdm_reference}(cdm, name) <- value
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{Name where to assign the new table.}

\item{value}{Table with the same source than the cdm object.}
}
\value{
The cdm reference.
}
\description{
Assign an table to a cdm reference.
}
