% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{summarisedResult}
\alias{summarisedResult}
\title{'summarised_results' object constructor}
\usage{
summarisedResult(x, name = "summarised_result")
}
\arguments{
\item{x}{input must be a tibble.}

\item{name}{Name of the summarised result object}
}
\value{
A summarisedResult object
}
\description{
'summarised_results' object constructor
}
