% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary.cdm_reference}
\alias{summary.cdm_reference}
\title{Summary a cdm reference}
\usage{
\method{summary}{cdm_reference}(object, ...)
}
\arguments{
\item{object}{A cdm reference object.}

\item{...}{For compatibility (not used).}
}
\value{
A summarised_result object with a summary of the data contained in
the cdm.
}
\description{
Summary a cdm reference
}
