% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary.cohort_table}
\alias{summary.cohort_table}
\title{Summary a generated cohort set}
\usage{
\method{summary}{cohort_table}(object, ...)
}
\arguments{
\item{object}{A generated cohort set object.}

\item{...}{For compatibility (not used).}
}
\value{
A summarised_result object with a summary of a cohort_table.
}
\description{
Summary a generated cohort set
}
