% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.summarised_result}
\alias{summary.summarised_result}
\title{Summary a summarised_result}
\usage{
\method{summary}{summarised_result}(object, ...)
}
\arguments{
\item{object}{A summarised_result object.}

\item{...}{For compatibility (not used).}
}
\value{
A summary of the result_types contained in a summarised_result
object.
}
\description{
Summary a summarised_result
}
\examples{
library(dplyr, warn.conflicts = FALSE)

person <- tibble(
  person_id = 1L, gender_concept_id = 0L, year_of_birth = 1990L,
  race_concept_id = 0L, ethnicity_concept_id = 0,
  month_of_birth = NA_integer_, day_of_birth = NA_integer_,
  birth_datetime = as.Date(NA_character_), location_id = 0L,
  provider_id = 0L, care_site_id = 0L, person_source_value = NA_character_,
  gender_source_value = NA_character_,
  gender_source_concept_id = NA_integer_, race_source_value = NA_character_,
  race_source_concept_id = NA_integer_,
  ethnicity_source_value = NA_character_,
  ethnicity_source_concept_id = NA_integer_
)
observation_period <- tibble(
  observation_period_id = 1L, person_id = 1L,
  observation_period_start_date = as.Date("2000-01-01"),
  observation_period_end_date = as.Date("2023-12-31"),
  period_type_concept_id = 0L
)
cdm <- cdmFromTables(
  tables = list("person" = person, "observation_period" = observation_period),
  cdmName = "test"
)

result <- summary(cdm)

summary(result)

}
