% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{suppress}
\alias{suppress}
\title{Function to suppress counts in summarised_result objects}
\usage{
suppress(
  result,
  minCellCount,
  variable,
  estimateType,
  group,
  groupCount,
  linkEstimates
)
}
\arguments{
\item{result}{summarised_result object}

\item{minCellCount}{Minimum count of records to report results.}

\item{variable}{Columns to suppress counts.}

\item{estimateType}{Types to suppress counts.}

\item{group}{Columns to group by and suppress.}

\item{groupCount}{Identifier of the group count.}

\item{linkEstimates}{List of estimates to link.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in summarised_result objects
}
