% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{suppress}
\alias{suppress}
\title{Function to suppress counts in summarised_result objects}
\usage{
suppress(result, minCellCount = 5)
}
\arguments{
\item{result}{Object to suppres counts}

\item{minCellCount}{Minimum count of records to report results.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in summarised_result objects
}
