% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSuppress.R
\name{suppress.summarised_result}
\alias{suppress.summarised_result}
\title{Function to suppress counts in summarised_result objects}
\usage{
\method{suppress}{summarised_result}(result, minCellCount = 5)
}
\arguments{
\item{result}{summarised_result object}

\item{minCellCount}{Minimum count of records to report results.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in summarised_result objects
}
