% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{tableName}
\alias{tableName}
\title{Read the table name of a \code{cdm_table}.}
\usage{
tableName(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\value{
A character with the name.
}
\description{
Read the table name of a \code{cdm_table}.
}
