% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{tableSource}
\alias{tableSource}
\title{Get the table source of a \code{cdm_table}.}
\usage{
tableSource(table)
}
\arguments{
\item{table}{A cdm_table.}
}
\value{
A cdm_source object.
}
\description{
Get the table source of a \code{cdm_table}.
}
