% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Convert a character to camelCase}
\usage{
toCamelCase(string)
}
\arguments{
\item{string}{String to convert to camelCase}
}
\description{
Convert a character to camelCase
}
\examples{
\donttest{
library(CDMUtilities)
toCamelCase(c("foo bar", "fooBar", "FooBar", "foo_bar", "FOOBAR"))
}

}
