% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/utilities.R
\name{to_snake_case}
\alias{to_snake_case}
\alias{toSnakeCase}
\title{Convert a character to snake_case}
\usage{
to_snake_case(string)

toSnakeCase(string)
}
\arguments{
\item{string}{String to convert to snake_case}
}
\description{
Convert a character to snake_case
}
\examples{
\donttest{
library(OMOPUtilities)
toSnakeCase(c("foo bar", "fooBar", "FooBar", "foo_bar", "FOOBAR"))
}

}
