% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{validateAchillesTable}
\alias{validateAchillesTable}
\title{Validate if a cdm_table is a valid achilles table.}
\usage{
validateAchillesTable(
  table,
  version = NULL,
  cast = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{table}{A cdm_table to validate.}

\item{version}{The cdm vocabulary version.}

\item{cast}{Whether to cast columns to required type.}

\item{call}{Passed to cli call.}
}
\value{
invisible achilles table
}
\description{
Validate if a cdm_table is a valid achilles table.
}
