% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAchillesTable.R
\name{validateAchillesTable}
\alias{validateAchillesTable}
\title{Validate if a cdm_table is a valid achilles table.}
\usage{
validateAchillesTable(
  table,
  version = NULL,
  cast = FALSE,
  empty = TRUE,
  nm = deparse1(substitute(table), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{table}{A cdm_table to validate.}

\item{version}{Version of the OMOP Common Data Model.}

\item{cast}{Whether to cast columns to the correct type.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{table}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
invisible achilles table
}
\description{
Validate if a cdm_table is a valid achilles table.
}
