% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateAgeGroupArgument}
\alias{validateAgeGroupArgument}
\title{validateAgeGroupArgument}
\usage{
validateAgeGroupArgument(ageGroup, overlap = FALSE, call = parent.frame())
}
\arguments{
\item{ageGroup}{age group in a list}

\item{overlap}{allow overlapping ageGroup}

\item{call}{parent frame}
}
\value{
validate ageGroup
}
\description{
validateAgeGroupArgument
}
