% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateAgeGroupArgument}
\alias{validateAgeGroupArgument}
\title{validateAgeGroupArgument}
\usage{
validateAgeGroupArgument(
  ageGroup,
  multipleAgeGroup = TRUE,
  overlap = FALSE,
  null = TRUE,
  call = parent.frame()
)
}
\arguments{
\item{ageGroup}{age group in a list}

\item{multipleAgeGroup}{allow mutliple age group}

\item{overlap}{allow overlapping ageGroup}

\item{null}{null age group allowed true or false}

\item{call}{parent frame}
}
\value{
validate ageGroup
}
\description{
validateAgeGroupArgument
}
