% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateCdmArgument}
\alias{validateCdmArgument}
\title{validateCdmArgument}
\usage{
validateCdmArgument(
  cdm,
  checkClass = TRUE,
  checkOverlapObservation = TRUE,
  checkStartBeforeEndObservation = TRUE,
  validation = "error",
  call = parent.frame()
)
}
\arguments{
\item{cdm}{A cdm_reference object}

\item{checkClass}{TRUE to check the cdm class}

\item{checkOverlapObservation}{TRUE to perform check on no overlap observation period}

\item{checkStartBeforeEndObservation}{TRUE to perform check on correct observational start and end date}

\item{validation}{How to perform validation: "error", "warning".}

\item{call}{A call argument to pass to cli functions.}
}
\value{
A cdm_reference object
}
\description{
validateCdmArgument
}
