% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGeneratedCohortSet.R
\name{validateCdmCohort}
\alias{validateCdmCohort}
\title{To assess if an object is a valid \code{cdm_cohort}}
\usage{
validateCdmCohort(cohortTable)
}
\arguments{
\item{cohortTable}{Table to validate}
}
\description{
To assess if an object is a valid \code{cdm_cohort}
}
