% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R, R/classCdmCohort.R
\name{validate_cdm_cohort}
\alias{validate_cdm_cohort}
\alias{validateCdmCohort}
\title{To assess if an object is a valid \code{omop_cohort}}
\usage{
validate_cdm_cohort(cohort_table)

validateCdmCohort(cohortTable)
}
\arguments{
\item{cohortTable}{\code{omop_cohort} to validate.}
}
\description{
To assess if an object is a valid \code{omop_cohort}
}
