% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{validateCdmReference}
\alias{validateCdmReference}
\title{Validate a \code{cdm_reference} object.}
\usage{
validateCdmReference(cdm)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}
}
\description{
Validate a \code{cdm_reference} object.
}
