% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmReference.R
\name{validateCdmReference}
\alias{validateCdmReference}
\title{Validate a CdmReferenceObject.}
\usage{
validateCdmReference(cdm)
}
\arguments{
\item{cdm}{A CdmReference object.}
}
\description{
Validate a CdmReferenceObject.
}
