% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCdmTable.R
\name{validateCdmTable}
\alias{validateCdmTable}
\title{Validate if a table is a valid cdm_table object.}
\usage{
validateCdmTable(table, name = NULL, call = parent.frame())
}
\arguments{
\item{table}{Object to validate.}

\item{name}{If we want to validate that the table has a specific name.}

\item{call}{Call argument that will be passed to \code{cli}.}
}
\value{
The table or an error message.
}
\description{
Validate if a table is a valid cdm_table object.
}
