% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateCohort}
\alias{validateCohort}
\title{Validate a cohort table input.}
\usage{
validateCohort(cohort)
}
\arguments{
\item{cohort}{Object to be validated as a valid cohort input.}
}
\description{
Validate a cohort table input.
}
