% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateCohortId}
\alias{validateCohortId}
\title{Validate A cohortId input.}
\usage{
validateCohortId(cohortId, cohort)
}
\arguments{
\item{cohortId}{A cohortId vector to be validated.}

\item{cohort}{A cohort_table object.}
}
\description{
Validate A cohortId input.
}
