% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateCohortIdArgument}
\alias{validateCohortIdArgument}
\title{Validate cohortId argument. CohortId can either be a cohort_definition_id
value, a cohort_name or a tidyselect expression referinc to cohort_names. If
you want to support tidyselect expressions please use the function as:
\code{validateCohortIdArgument({{cohortId}}, cohort)}.}
\usage{
validateCohortIdArgument(
  cohortId,
  cohort,
  validation = "error",
  call = parent.frame()
)
}
\arguments{
\item{cohortId}{A cohortId vector to be validated.}

\item{cohort}{A cohort_table object.}

\item{validation}{How to perform validation: "error", "warning".}

\item{call}{A call argument to pass to cli functions.}
}
\description{
Validate cohortId argument. CohortId can either be a cohort_definition_id
value, a cohort_name or a tidyselect expression referinc to cohort_names. If
you want to support tidyselect expressions please use the function as:
\code{validateCohortIdArgument({{cohortId}}, cohort)}.
}
