% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateName}
\alias{validateName}
\title{Validate that the a variable is a valid name for a table in the cdm. A
message will be shown if name had to be modified and/or the name is already
part of the cdm.}
\usage{
validateName(name, cdm = NULL)
}
\arguments{
\item{name}{Name of a new table to be added to a cdm object.}

\item{cdm}{A cdm_reference object. It will tro}
}
\description{
Validate that the a variable is a valid name for a table in the cdm. A
message will be shown if name had to be modified and/or the name is already
part of the cdm.
}
