% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateNameArgument}
\alias{validateNameArgument}
\title{Validate name argument. It must be a snake_case character vector. You can add
a cdm object to check that \code{name} is not already used in that cdm.}
\usage{
validateNameArgument(
  name,
  cdm = NULL,
  validation = "error",
  null = FALSE,
  empty = TRUE,
  nm = deparse1(substitute(name), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{name}{Name of a new table to be added to a cdm object.}

\item{cdm}{A \verb{<cdm_reference>} object. If provided, the name will be checked
to ensure it does not already exist in the cdm.}

\item{validation}{How to perform validation: "error", "warning".}

\item{null}{If TRUE, name can be NULL}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{name}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\description{
Validate name argument. It must be a snake_case character vector. You can add
a cdm object to check that \code{name} is not already used in that cdm.
}
\examples{
# this is a validate name
name <- "my_new_table"
validateNameArgument(name)

# this is not
name <- "myTableNAME"
validateNameArgument(name, validation = "warning")

}
