% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateNameArgument}
\alias{validateNameArgument}
\title{Validate name argument.}
\usage{
validateNameArgument(
  name,
  cdm = NULL,
  validation = "error",
  call = parent.frame()
)
}
\arguments{
\item{name}{Name of a new table to be added to a cdm object.}

\item{cdm}{A cdm_reference object. It will check if a table named name
already exists in the cdm.}

\item{validation}{How to perform validation: "error", "warning".}

\item{call}{A call argument to pass to cli functions.}
}
\description{
Validate name argument.
}
