% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSummarisedResult.R
\name{validateNameLevel}
\alias{validateNameLevel}
\title{Validate if two columns are valid Name-Level pair.}
\usage{
validateNameLevel(
  x,
  prefix,
  validation = "error",
  empty = TRUE,
  nm = deparse1(substitute(x), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{x}{A tibble.}

\item{prefix}{Prefix for the name-level pair, e.g. 'strata' for
strata_name-strata_level pair.}

\item{validation}{Either 'error', 'warning' or 'message'.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{x}.}

\item{call}{Will be used by cli to report errors.}
}
\description{
Validate if two columns are valid Name-Level pair.
}
