% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateNewColumn}
\alias{validateNewColumn}
\title{Validate a new column of a table}
\usage{
validateNewColumn(
  table,
  column,
  empty = TRUE,
  validation = "warning",
  nm = deparse1(substitute(column), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{table}{The table to check if the column already exists.}

\item{column}{Character vector with the name(s) of the new column(s).}

\item{empty}{Whether it can be empty.}

\item{validation}{How to perform validation: "error", "warning".}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{column}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
table without conflicting columns.
}
\description{
Validate a new column of a table
}
\examples{
x <- dplyr::tibble(
  column1 = c(1L, 2L),
  column2 = c("a", "b")
)
validateNewColumn(x, "not_exiting_column")
validateNewColumn(x, "column1")

}
