% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOmopTable.R
\name{validateOmopTable}
\alias{validateOmopTable}
\title{Validate an omop_table}
\usage{
validateOmopTable(
  omopTable,
  version = NULL,
  cast = FALSE,
  empty = TRUE,
  nm = deparse1(substitute(omopTable), backtick = TRUE),
  call = parent.frame()
)
}
\arguments{
\item{omopTable}{An omop_table to check.}

\item{version}{Version of the OMOP Common Data Model. Supported options are:
"5.3", "5.4", or "5.5".}

\item{cast}{Whether to cast columns to the correct type.}

\item{empty}{Whether it can be empty.}

\item{nm}{Name to use in error messages. Defaults to the expression supplied
to \code{omopTable}.}

\item{call}{Call argument passed to \code{cli} functions.}
}
\value{
An omop_table object.
}
\description{
Validate an omop_table
}
