% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateResultArguemnt}
\alias{validateResultArguemnt}
\title{validateResultArguemnt}
\usage{
validateResultArguemnt(result, validation = "error", call = parent.frame())
}
\arguments{
\item{result}{summarise result object to validate}

\item{validation}{message to return}

\item{call}{parent.frame}
}
\value{
summarise result object
}
\description{
validateResultArguemnt
}
