% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateResultArgument}
\alias{validateResultArgument}
\title{validateResultArgument}
\usage{
validateResultArgument(
  result,
  checkNoDuplicates = FALSE,
  checkNameLevel = FALSE,
  checkSuppression = FALSE,
  validation = "error",
  call = parent.frame()
)
}
\arguments{
\item{result}{summarised_result object to validate.}

\item{checkNoDuplicates}{Whether there are not allowed duplicates in the
result object.}

\item{checkNameLevel}{Whether the name-level paired columns are can be
correctly split.}

\item{checkSuppression}{Whether the suppression in the result object is well
defined.}

\item{validation}{Only error is supported at the moment.}

\item{call}{parent.frame}
}
\value{
summarise result object
}
\description{
validateResultArgument
}
