% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateWindowArgument}
\alias{validateWindowArgument}
\title{validateWindowArgument}
\usage{
validateWindowArgument(window, validation = "error", call = parent.frame())
}
\arguments{
\item{window}{time window}

\item{validation}{How to perform validation: "error", "warning".}

\item{call}{A call argument to pass to cli functions.}
}
\value{
time window
}
\description{
validateWindowArgument
}
