% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{writeLogger}
\alias{writeLogger}
\title{Write message to a logger file associated to a cdm object.}
\usage{
writeLogger(message, warn = FALSE)
}
\arguments{
\item{message}{Message to record in the logger.}

\item{warn}{Weather to throw a warning if the logger does not exist.}

\item{cdm}{A cdm reference.}
}
\value{
Invisible cdm object
}
\description{
Write message to a logger file associated to a cdm object.
}
